% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_3b_HCRs.R, R/MP_3c_HCR_ICES_MSY.R
\name{annualTAC}
\alias{annualTAC}
\alias{IcesHCR}
\title{Harvest Control Rules (HCRs)}
\usage{
annualTAC(stocks, advice, advice.ctrl, year, stknm, ...)

IcesHCR(stocks, advice, advice.ctrl, year, stknm, ...)
}
\arguments{
\item{stocks}{And FLStocks object.}

\item{advice}{A list with two FLQuant elements, TAC and quota.share. TAC is an FLQuant  with quant dimension equal to the number of stocks in biols object, the names used in in the quant dimension must be equal to those used in biols. quota.share is a list with one element per stock in biols object indicating the quota share per stock and fleet. The quant dimension of the elements must be equal to the number of fleets and the names used must be equal to those in fleets objects.}

\item{advice.ctrl}{A list with the settings to control the advice model for each stock (the HCR for each stock, the reference points used in the HCR, additional parameters, ...)}

\item{year}{The position of the assessment year in the stocks and advice objects.}

\item{stknm}{The name of the stock for which advice is being generated.}

\item{...}{Any extra arguments needed for specific HCRs.}
}
\value{
The advice input object updated with the management advice (TAC) generated by the HCR.
}
\description{
There are several HCRs available in FLBEIA which are used within the main function FLBEIA to generate the management advice in each step.
But they can also be used independently.
}
\details{
There are two types of HCRs model-free HCRs and model-Based HCRs. 
         Model-free HCRs use abundance indices to generate the advice and hence it use  FLIndices object as input data.
         Model-based HCRs use estimates of stock abundance and stock exploitation level to generate the advice.
\itemize{
     \item    Model-Free HCRs: annexIVHCR and ghlHCR.
     \item    Model-Based HCRs: CFPMSYHCR, F2CatchHCR, FroeseHCR, IcesHCR, MAPHCR, annualTAC, neaMAC_ltmp, pilRec_ane, aneHCRE, 
                          aneRec_pil, neaMAC_ltmp, pilRec_ane, aneHCRE and aneRec_pil.
}
\itemize{
      \item{annexIVHCR}{is the HCR used by EC and ICES to generate the TAC advice for data poor stocks}
      \item{ghlHCR}{Is a HCR used in the management of greenland-halibut}
      \item{CFPMSYHCR} 
      \item{FroeseHCR} {The HCR defined in the paper by Froese, Branch et al. in Fish and Fisheries 2010.}
      \item{IcesHCR} {The HCR used by ICES to generate TAC advice in the MSY framework.}
      \item{MAPHRC} {The HCR proposed by the EC in the evaluation on multi-annual management plans in 2015.}
      \item{annualTAC} {A HCR that generates annual TAC advice. The HCR provides the whole flexibility of fwd }
}       
        The HCRs are documented in the manual of the library.
}
