% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions_biolfleets2flstock.R
\name{biolfleets2flstock}
\alias{biolfleets2flstock}
\title{Function to generate an FLStock from an FLBiol and FLFleet}
\usage{
biolfleets2flstock(biol, fleets)
}
\arguments{
\item{biol}{An FLBiolobject}

\item{fleets}{An FLFleetExt object}
}
\value{
An FLStock object with abundance and biological information from biol argument 
        and exploitation levels from fleets argument
}
\description{
Function to generate an FLStock from an FLBiol and FLFleet
}
\examples{
\dontrun{

data(res_flbeia)

biol   <- oneRes$biols$stk1
fleets <- oneRes$fleets

stk <- biolfleets2flstock( biol=biol, fleets=fleets)

}
}
