% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Summary.R
\name{bioSum}
\alias{bioSum}
\title{Summary of the FLBEIA output}
\usage{
bioSum(obj, stocks = "all", years = dimnames(obj$biols[[1]]@n)$year,
  long = TRUE, scenario = "bc")
}
\arguments{
\item{obj}{The output of the FLBEIA function.}

\item{years}{The period of years.}

\item{long}{The data frame should be constructed using long or wide format?}

\item{flnms}{Names of the fleets.}

\item{stknms}{Names of the stocks.}
}
\value{
The data frames can be of wide or long format. In long format all the indicators are in the same column.
 There is one column, indicator, for the name of the indicator and a second one value for the numerica value of the indicator.
 In the wide format each of the indicators correspond with one column in the data frame. 
 The long format it is recommendable to work with ggplot2 functions for example and the wide format 
 it is more efficient for memory allocation and speed of computations.

A data frame.
}
\description{
Summarize the results of the simulation in data frames.
}
\details{
\itemize{
     \item{bioSum:} Data frame with the biological indicators The columns are stocks, 
     years, it, indicators and value. 
     \item{fleetSum:} The colums of the data frame are: year, quarter, stock, fleet, iter, profits, 
     capacity, costs, discards, effort and landings.
     \item{ecoSum_damara:} ecoSum for Damara project.
     \item{metierSum:} The colums of the data frame are: year, quarter, fleet, metier, iter, 
     effort and effshare.
     \item{stockFleetSum:} The colums of the data frame are: year, quarter, stock, fleet, iter, 
     landings, discards, price and tacshare.
     \item{stockMetierSum:} The colums of the data frame are: year, quarter, stock, fleet, metier, iter, 
     landings, discards and price.
     
}
}
\examples{
\dontrun{

library(FLBEIA)
library(ggplot2)

# Apply the summary functions to the examples runs in FLBEIA help page.
# Test the different arguments in summary function.

data(res_flbeia)

# Example One: One stock, one fleet, one iter.
s0_bio    <- bioSum(s0)
s0_flt    <- fltSum(s0)

s0_bioQ   <- bioSumQ(s0_bio)
s0_fltQ   <- fltSumQ(s0_flt)

s0_bio    <- bioSum(s0, long = FALSE, years = ac(2016:2020))
s0_flt    <- fltSum(s0, long = FALSE, years = ac(2016:2020))

s0_bioQ   <- bioSumQ(s0_bio)
s0_fltQ   <- fltSumQ(s0_flt)

s0_flt    <- fltSum(s0, long = FALSE, byyear = FALSE)
s0_fltQ   <- fltSumQ(s0_flt)


# Example OneIters: As one but with iterations.
s1_bio    <- bioSum(s1, years = ac(2016:2020))
s1_flt    <- fltSum(s1, years = ac(2016:2020))

s1_bioQ   <- bioSumQ(s1_bio)
s1_fltQ   <- fltSumQ(s1_flt)

s1_bio    <- bioSum(s1, long = FALSE)
s1_flt    <- fltSum(s1, long = FALSE)

s1_bioQ   <- bioSumQ(s1_bio)
s1_fltQ   <- fltSumQ(s1_flt)

s1_flt    <- fltSum(s1, long = FALSE, byyear = FALSE)
s1_fltQ   <- fltSumQ(s1_flt)

# Example Multi: Two stock, two fleet, four iters. 
s2_bio    <- bioSum(s2, years = ac(2016:2020))
s2_flt    <- fltSum(s2, years = ac(2016:2020))

s2_bioQ   <- bioSumQ(s2_bio)
s2_fltQ   <- fltSumQ(s2_flt)

s2_bio    <- bioSum(s2, long = FALSE)
s2_flt    <- fltSum(s2, long = FALSE)

s2_bioQ   <- bioSumQ(s2_bio)
s2_fltQ   <- fltSumQ(s2_flt)

s2_flt    <- fltSum(s2, long = FALSE, byyear = FALSE)
s2_fltQ   <- fltSumQ(s2_flt)

s2_bio    <- bioSum(s2, stocks = 'stk2')
s2_flt    <- fltSum(s2, flnms = 'fl1')

s2_bioQ   <- bioSumQ(s2_bio)
s2_fltQ   <- fltSumQ(s2_flt)

}
}
