% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brps_seasonal.R
\name{brpsson}
\alias{brpsson}
\title{Reference points for a seasonal model}
\usage{
brpsson(
  stk,
  B0,
  R0,
  rec.ss = 1,
  ssb.ss = 1,
  sr_model,
  sr_params,
  Fprop = rep(1/dim(stk)[4], dim(stk)[4]),
  Fscan = seq(0, 4, by = 0.01),
  oldest = 100,
  nrun = 200,
  tol = 0.01
)
}
\arguments{
\item{stk}{An FLStock object.}

\item{B0}{The value of the virgin biomass.}

\item{R0}{The expected recrutiment in the virgin population.}

\item{rec.ss}{The recruitment season (numeric). Default value = 1.}

\item{ssb.ss}{The spawning season (numeric). Default value = 1.}

\item{sr_model}{A character with the name of the model to simulate the recruitment process.}

\item{sr_params}{A named vector with the SR parameter values.}

\item{Fprop}{A vector with the same length as the number of seasons with the proportion of F by season. 
By default the same proportion in all the seasons is assumed.}

\item{Fscan}{A vector with the F values to be simulated. 
By default, will scan values between 0 and 4 (by 0.01 increments).}

\item{oldest}{The maximum age class to be considered. Default 100.}

\item{nrun}{The maximum number of years to project forward until equilibrium.}

\item{tol}{The desired accuracy.}
}
\value{
A list with 2 elements runs and refpts

A list with two elements:
\itemize{
  \item{\code{runs}}{A data.frame containing a summary of the simulation
  results (year by year until equilibrium is reached): target F (ftgt),
  initial proportion of F in each season (fprop), expected biomass, B1+,
  SSB, recruitment, spawning biomass per recruit (ssbpr), global fishing
  mortality (fbar), global F proportion by season (pF_s), expected yields
  and their seasonal proportions (pY_s), yield per recruit (YPR), harvest
  rates (hr_b1plus = catch/b1plus and hr_ssb = catch/ssb), convergence
  status (TRUE/FALSE), virgin status levels (rec0, ssb0, ssbpr0), status
  relative to virgin levels (ratioR, ratioB, ratioSPR), slope (MSY:
  slope = 0 at equilibrium) and rslope (F0.1: rslope = 0.1 at equilibrium).}

  \item{\code{refpts}}{A data.frame containing the estimated reference
  points and equilibrium values for selected variables (ftgt, fprop_s1,
  fprop_s2, biomass, b1plus, ssb, rec, ssbpr, fbar). Reported reference
  levels include MSY, F0.1, \%SPR (spr.p), \%B0 (F.pB0) for different
  percentages, and the fishing mortality leading to 50\%R0 (F.50R0).}
}
}
\description{
Function to estimate reference points when having a seasonal model
}
\examples{

library(FLBEIA)
data(multistk) # object with 2 seasons and 3 iterations

stk <- trim( multistk, year=1, iter = 1) 

# loop for different catch proportions by season
for (p in seq(0.1,0.9,0.1)) {
  fruns <- brpsson( stk, B0=1e+05, R0=27489766, rec.ss=2, ssb.ss=2, 
                    sr_model="bevholt", sr_params=c( a = 29988835.109, b = 9090.909), 
                    Fprop = c(p, 1-p))
} 

plotBRPsson( fruns, pdfnm="stk_Fbar_vs_SPR.pdf")
 
}
\seealso{
\code{\link{plotBRPsson}}
}
