% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_calculate.CDparam.R
\name{calculate.q.sel.flrObjs}
\alias{calculate.q.sel.flrObjs}
\title{Calculates selectivity-related parameters}
\usage{
calculate.q.sel.flrObjs(biols, fleets, BDs, fleets.ctrl, mean.yrs, sim.yrs)
}
\arguments{
\item{biols}{An FLBiols object.}

\item{fleets}{An FLFleetsExt object. An extended version of the FLFleet object defined in FLCore.}

\item{BDs}{A list of FLSRsim objects. One per biomass dynamic stock in biols object.}

\item{fleets.ctrl}{Fleets' control file containing the catch production function for each fleet and stock.}

\item{mean.yrs}{A character vector with the name of the years used to calculate mean selectivity.}

\item{sim.yrs}{A character vector with the name of the years in the projection period.}
}
\value{
A FLFleetsExt object.
}
\description{
Given stock abundances and catches (i.e. landings and discards), this function estimates values for
\code{fleets[[fl]]@metiers[[mt]]@catches[[st]]@catch.q}, 
\code{fleets[[fl]]@metiers[[mt]]@catches[[st]]@landings.sel}, and 
\code{fleets[[fl]]@metiers[[mt]]@catches[[st]]@discards.sel} 
for all years except the simulation years. For the simulation years, 
the parameter values are set as the mean of the parameters along \code{mean.yrs}.
}
