% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Checks.R
\name{checkBiols}
\alias{checkBiols}
\alias{checkFleets}
\alias{checkSRs}
\alias{checkAdvice}
\alias{checkObsctrl}
\title{Checking arguments of FLBEIA function}
\usage{
checkBiols(object)

checkFleets(object, ctrl = NULL)

checkSRs(object)

checkAdvice(object)

checkObsctrl(object)
}
\arguments{
\item{object}{An object of the appropiate class depending on the FLBEIA argument to be checked:
- biols   : A FLBiols object.
- fleets  : A FLFleetsExt object.
- SRs     : A list of FLSRsim objects.
- advice  : A list with two FLQuant elements, TAC and quota.share.
- obs.ctrl: A list.}
}
\value{
An error message if there is any error detected.
}
\description{
This functions checks wether some conditions are met for the inputs 
             in the different arguments of the FLBEIA function.
             Specifically there are checking functions for: 
             biols, fleets, SRs, advice, and obs.ctrl arguments.

Functions available:
\itemize{

   \item{\code{checkBiols}}{: for checking biols argument of FLBEIA function 
                              (of class FLBiols).}
   \item{\code{checkFleets}}{: for checking fleets argument of FLBEIA function 
                              (of class FLFleetsExt).}
   \item{\code{checkSRs}}{: for checking SRs argument of FLBEIA function 
                              (list of FLSRsim objects).}
   \item{\code{checkAdvice}}{: for checking advice argument of FLBEIA function 
                              (of class list with two FLQuant elements, TAC and quota.share).}
   \item{\code{checkObsctrl}}{: for checking obs.ctrl argument of FLBEIA function 
                              (of class list).}
}
}
