% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Checks.R
\name{checkFLBEIAData}
\alias{checkFLBEIAData}
\alias{checkBiols}
\alias{checkFleets}
\alias{checkSRs}
\alias{checkBDs}
\alias{checkAdvice}
\alias{checkObsctrl}
\title{Checking inputs of FLBEIA function}
\usage{
checkFLBEIAData(
  biols,
  SRs = NULL,
  BDs = NULL,
  fleets,
  covars = NULL,
  indices = NULL,
  advice = NULL,
  main.ctrl,
  biols.ctrl,
  fleets.ctrl,
  covars.ctrl,
  obs.ctrl,
  assess.ctrl,
  advice.ctrl
)

checkBiols(object)

checkFleets(object, ctrl = NULL)

checkSRs(object)

checkBDs(object)

checkAdvice(object)

checkObsctrl(object)
}
\arguments{
\item{biols}{An FLBiols object.}

\item{SRs}{A list of FLSRsim objects.}

\item{BDs}{A list of FLBDsim objects.}

\item{fleets}{An FLFleetsExt object.}

\item{covars}{A list of FLQuants.}

\item{indices}{A list of FLIndices.}

\item{advice}{A list with two FLQuant elements, TAC and quota.share.}

\item{main.ctrl}{A list with the settings to control the main function.}

\item{biols.ctrl}{A list with the settings to control the biological operating model for each stock.}

\item{fleets.ctrl}{A list with the settings to control the fleets operating model for each fleet.}

\item{covars.ctrl}{A list with the settings to control the covars operating model for each fleet.}

\item{obs.ctrl}{A list with the settings to control the observation model for each stock.}

\item{assess.ctrl}{A list with the settings to control the specify the assessment model for each stock.}

\item{advice.ctrl}{A list with the settings to control the advice model for each stock.}

\item{object}{An object of the appropiate class depending on the FLBEIA argument to be checked:
\itemize{
  \item{\code{biols}}   {: A FLBiols object.}
  \item{\code{fleets}}  {:A FLFleetsExt object.}
  \item{\code{SRs}}     {: A list of FLSRsim objects.}
  \item{\code{BDs}}     {: A list of FLBDsim objects.}
  \item{\code{advice}}  {: A list with two FLQuant elements, TAC and quota.share..}
  \item{\code{obs.ctrl}}{: A list.}
}}

\item{ctrl}{An object of class list (optional argument), input for fleets.ctrl argument of FLBEIA function.}
}
\value{
An error message if there is any error detected.
}
\description{
This functions checks wether some conditions are met for the inputs 
             in the different arguments of the FLBEIA function.
             There are also checking functions for specific inputs: 
             biols, fleets, SRs, advice, and obs.ctrl arguments.

Functions available:
\itemize{

   \item{\code{checkFLBEIAData}}{: for checking inputs for differents arguments of FLBEIA function 
                              (of class FLBiols).}
    
   \item{\code{checkBiols}}{: for checking biols argument of FLBEIA function 
                              (of class FLBiols).}
   \item{\code{checkFleets}}{: for checking fleets argument of FLBEIA function 
                              (of class FLFleetsExt).}
   \item{\code{checkSRs}}{: for checking SRs argument of FLBEIA function 
                              (list of FLSRsim objects).}
   \item{\code{checkAdvice}}{: for checking advice argument of FLBEIA function 
                              (of class list with two FLQuant elements, TAC and quota.share).}
   \item{\code{checkObsctrl}}{: for checking obs.ctrl argument of FLBEIA function 
                              (of class list).}
}
}
