% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_CtrlCreators_advice.r
\name{create.advice.ctrl}
\alias{create.advice.ctrl}
\title{advice.ctrl object creator}
\usage{
create.advice.ctrl(stksnames, HCR.models = NULL, ...)
}
\arguments{
\item{stksnames}{A vector with the name of the stocks in the OM.}

\item{HCR.models}{A character vector of the same length as stksnames with the name of the HCR used to generate the management advice.}

\item{...}{any extra arguments necessary in the HCR specific creators. '...' are extracted using 'list(...)', this generates a named list with the extra arguments.
To assure the correct functioning the extra arguments must have a name.}

\item{first.year}{The first year in which advice is calculated.}

\item{last.year}{The last year in which advice is calculated.}
}
\value{
A list of lists with the basic structure of the advice.ctrl object.
}
\description{
It creates the advice.ctrl object to be used in the call to the main function FLBEIA.
}
