% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_CtrlCreators_assessment.r
\name{create.assess.ctrl}
\alias{create.assess.ctrl}
\title{assess.ctrl object creator}
\usage{
create.assess.ctrl(stksnames, assess.models = NULL, assess.ctrls = NULL,
  ...)
}
\arguments{
\item{stksnames}{A vector with the name of the stocks in the OM.}

\item{assess.models}{A character vector of the same length as stksnames with the name of the model used to obtaine the perceived population in the MP.}

\item{assess.ctrls}{A list of the same length as stksnames with the arguments needed to fit the assessment model.}

\item{...}{any extra arguments necessary in the model specific creators. '...' are extracted using 'list(...)', this generates a named list with the extra arguments.
To assure the correct functioning the extra arguments must have a name.}
}
\value{
A list of lists with the basic structure of the assess.ctrl object.
}
\description{
It creates the assess.ctrl object to be used in the call to the main function FLBEIA.
}
