% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_create.BDs.data.R
\name{create.BDs.data}
\alias{create.BDs.data}
\title{Run create.fleets.data create a FLFleetsExt object}
\usage{
create.BDs.data(yrs, ns, ni, stks.data)
}
\arguments{
\item{yrs}{A vector with c(first.yr,proj.yr, last.yr) where
\itemize{
     \item first.yr: First year of simulation (number).
     \item proj.yr: First year of projection (number).
     \item last.yr: Last year of projection (number).}}

\item{ns}{Number of seasons (number).}

\item{ni}{Number of iterations (number).}
}
\description{
create.biols.data create an FLBiols object
}
