% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_CtrlCreators_biols.R
\name{create.biols.ctrl}
\alias{create.biols.ctrl}
\title{biols.ctrl object creator}
\usage{
create.biols.ctrl(stksnames, growth.models = NULL, immediate = FALSE, ...)
}
\arguments{
\item{stksnames}{A vector with the name of the stocks in the OM.}

\item{growth.models}{A character vector of the same length as stksnames with the name of the model used to project the stock populations in the simulation.}

\item{immediate}{logical, indicating if the warnings should be output immediately.}

\item{...}{any extra arguments necessary in the model specific creators. '...' are extracted using 'list(...)', this generates a named list with the extra arguments.
To assure the correct functioning the extra arguments must have a name.}
}
\value{
A list of lists with the basic structure of the biols.ctrl object.
}
\description{
It creates the biols.ctrl object to be used in the call to the main function FLBEIA.
}
