% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_create.biols.data.R
\name{create.biols.data}
\alias{create.biols.data}
\title{Run create.biols.data create an FLBiols object}
\usage{
create.biols.data(yrs, ns, ni, stks.data)
}
\arguments{
\item{yrs}{a vector with c(first.yr,proj.yr, last.yr) where
*          first.yr: First year of simulation (number)}

\item{ns}{Number of seasons (number)}

\item{ni}{Number of iterations (number)}

\item{proj.yr:}{First year of projection (number)}

\item{last.yr:}{Last year of projection (number)}

\item{stks.data:}{a list with the name of the stks and with the next elements}

\item{stks:}{Name of all the stocks (vector)}

\item{stk.unit}{Number of units of the stock (number)}

\item{stk.min.age:}{Minimum age of the stock (number)}

\item{stk.max.age:}{Maximum age of the stock (number)}

\item{stk_n.flq:}{Abundance at age age (FLQuant)}

\item{stk_wt.flq:}{Weight age age (FLQuant)}

\item{stk_m.flq:}{Natural mortality age age (FLQuant)}

\item{stk_fec.flq:}{Fecundity at age    (FLQuant)}

\item{stk_spwn.flq:}{Spawning at age    (FLQuant)}

\item{stk_range.min:}{Minimum age (number)}

\item{stk_range.max:}{Maximum age (number)}

\item{stk.range.minyear:}{Minimum year (number)}

\item{stk_range.minfbar:}{Minimum year to take into account in the calculation of 'f' (number)}

\item{stk_range.maxfbar:}{Maximum year to take into account in the calculation of 'f' (number)}

\item{stk_biol.proj.avg.yrs:}{historic years to calculate the average of spwn, fec, m and wt (vector)}
}
\description{
create.biols.data create an FLBiols object
}
