% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_CtrlCreators_covars.r
\name{create.covars.ctrl}
\alias{create.covars.ctrl}
\title{covars.ctrl object creator}
\usage{
create.covars.ctrl(
  cvrsnames,
  process.models = NULL,
  flq,
  immediate = FALSE,
  ...
)
}
\arguments{
\item{cvrsnames}{A vector with the name of the covariates in the OM.}

\item{process.models}{A character vector of the same length as cvrsnames with the name of the process model
followed by each of the covariates.
The first element corresponds with the process model of the first covariable in cvrsnames, 
the second with the second and so on. 
The default is NULL in which case 'fixedCovar' is used for **all** the covariates.}

\item{flq}{An FLQuant to give structure to the FLQuants to be used within the function, 
the dimension and dimnames in 'year', 'season' and 'iter' will be used to create the necessary FLQuants.}

\item{immediate}{logical, indicating if the warnings should be output immediately.}

\item{...}{any extra arguments necessary in the HCR specific creators. '...' are extracted using 'list(...)', this generates a named list with the extra arguments.
To assure the correct functioning the extra arguments must have a name.}
}
\value{
A list of lists with the basic structure of the covars.ctrl object.
}
\description{
It creates the covars.ctrl object to be used in the call to the main function FLBEIA.
}
