% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_create.ecoData.arrays.R
\name{create.ecoData}
\alias{create.ecoData}
\title{Function to generate the covar object neccesary to compute economic indicators and fills in economic slots in FLFleets object}
\usage{
create.ecoData(file, fltObj, hist.yrs, mean.yrs, sim.yrs)
}
\arguments{
\item{file}{An excel file with the economic data. The names used for the sheets and the columns must be the same used to name fleets and metiers in fltObj. However, the fleets in the excel file can be a subset of the fleets in the fleetsObj, i.e., is not neccesary to provide economic data for all the fleets. The order of metiers in the columns data must correspond with the name of the metiers used in the FLFleet object.}

\item{fltObj}{An FLFleets object  with the structure of the fleet and which may contain historical data.}

\item{hist.yrs}{A vector with the historical years.}

\item{mean.yrs}{A vector with the years used to compute the mean to condition the parameters in the projection period.}

\item{sim.yrs}{A vector with the simulation years.}
}
\value{
An /code{FLFleetsExt}.
}
\description{
This function generates an covar object with the economic indicators and fills in vcost, fcost, capacity and crewshare 
            slots in FLFleets object. The data is given in excel File (xls and xlsx).
}
\seealso{
\code{\link{FLFleetsExt}}, \code{\link{create.biol.arrays}}
}
\author{
Dorleta Garcia & Sonia Sanchez.
}
\keyword{create.fleets.arrays}
