% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_create.fleets.arrays.R
\name{create.fleets.arrays}
\alias{create.fleets.arrays}
\title{Function to generate an FLFleetsExt object given inputs as arrays}
\usage{
create.fleets.arrays(
  stk_objs,
  caa_objs,
  caa_objs_path,
  price_objs,
  price_objs_path,
  catch_obj,
  effort_obj,
  flt_obj = NULL,
  stk_nms = NA,
  flt_nms,
  flt_mt_nms,
  flt_mt_stk_nms,
  ages = NULL,
  hist.yrs,
  sim.yrs,
  mean.yrs,
  new_hist.yrs = hist.yrs,
  update_catch_effort = TRUE,
  update_price = TRUE,
  update_weight = TRUE,
  caa_flt_mt_correspondences = NULL,
  paa_flt_mt_correspondences = NULL,
  caaOpt,
  priceOpt,
  excel = TRUE
)
}
\arguments{
\item{stk_objs}{A character vector with the names of the files containing the stocks data. See \link{create.biol.arrays} for more detail.
Supported format is only Excel (xls and xlsx), each stock can be in different format.}

\item{caa_objs}{A character vector with the names of the files containing the catch at age data (in numbers), both for landings and discards.
Supported formats are Excel (xls and xlsx) and R format (RData), each file can be in different format.
The number of required files depend of the value of \code{caaOpt} argument:
\itemize{ 
   \item{\code{caaOpt} = 1 or 2   }{ => one per stock and fleet. Named vector stknm_fltnm}
   \item{\code{caaOpt} = 3, 4 or 5}{ => one per stock. Named vector stknm}
}
If NULL, the function looks for "caa_stknm_flnm.xlsx" in \code{caaOpt} = 1 or 2 and "caa_stknm.xlsx" in \code{caaOpt} = 3 ,4 or 5.}

\item{caa_objs_path}{A character vector with the \code{caa_objs} file path.}

\item{price_objs}{A character vector with the names of the files containing the price at age data.
Supported formats are Excel (xls and xlsx) and R format (RData), each prices file can be in different format.
The number of required files depend of the value of \code{caaOpt} argument:
\itemize{ 
   \item{\code{caaOpt} = 1 or 2   }{ => one per stock and fleet. Named vector stknm_fltnm.}
   \item{\code{caaOpt} = 3, 4 or 5}{ => one per stock. Named vector stknm.}
}}

\item{price_objs_path}{A character vector with the price_objs file path.}

\item{catch_obj}{A character vector with the names of the files containing the catch data in Fcube format.
Supported formats are Excel (xls and xlsx) and R format (RData) and required columns are 'year', 'fleet', 'metier', 'stock', 'category' and 'catch'.}

\item{effort_obj}{A character vector with the names of the files containing the effort data in Fcube format.
Supported formats are Excel (xls and xlsx) and R format (RData) and required columns are 'year', 'fleet', 'metier' and 'effort'.
Both \code{catch_obj} and \code{effort_obj} must be in the same format.}

\item{flt_obj}{An FLFleets object (optional) with the structure of the fleet and which may contain historical data. 
If this object is provided, then the arguments \code{stk_nms}, \code{flt_nms}, \code{flt_mt_nms} and \code{ages} will not be used
and \code{excel} argument must be set to \code{FALSE}.}

\item{stk_nms}{A character vector (optional) with the name of all the stocks caugth by the different fleets.}

\item{flt_nms}{A character vector with the name of the fleets.}

\item{flt_mt_nms}{A list with one element per fleet. In turn, each element is a character vector with the names of the metiers in the corresponding fleet.}

\item{flt_mt_stk_nms}{A list with one element per fleet and metier. In turn, each element is a character vector with the names of the stocks in the corresponding fleet and metier.}

\item{ages}{A list with one element per stock, with the age classes of the stock.}

\item{hist.yrs}{A vector with the historical years.}

\item{sim.yrs}{A vector with the simulation years.}

\item{mean.yrs}{A vector with the years used to compute the mean to condition the parameters in the projection period.}

\item{new_hist.yrs}{A vector with the years from input files that will be used to condition the parameters in the historic years.
If a value is not provided, the it is set equal to \code{hist.yrs}.}

\item{update_catch_effort}{Logical. If \code{TRUE} (default), catch and effort must be provided and \code{catch_obj}, \code{effort_obj} arguments are required.}

\item{update_price}{Logical. If \code{TRUE} (default), prices must be provided and \code{price_objs} and \code{price_objs_path} arguments are required.}

\item{update_weight}{Logical. If \code{TRUE} (default), weights at age for landings and discards must be provided, 
so wl and wd sheets are required in files listed in \code{stk_objs} argument.}

\item{caa_flt_mt_correspondences}{An Excel file name. This file must contain one sheet per stock, with the correspondences between the fleet segments used in \code{caaa_obj} data 
and the fleet metier segmentation used in the analysis. If the file does not exist, it is supposed that the caa data is given by fleet and metier.}

\item{paa_flt_mt_correspondences}{An Excel file name. This file must contain information on prices correspondences, with same format and requirements as \code{caa_flt_mt_correspondences} argument.}

\item{caaOpt}{A code number to determine the way in wich catch at age data are provided.
The option to be used depends on the data availabiltiy, from data rich to data-poor and the following codes are available:
\itemize{ 
   \item{1}{: If catch at age data is available at metier level for all the metiers.}
   \item{2}{: If catch at age data is only available at fleet level.}
   \item{3}{: If catch at age data is disaggregated but the segments do not correspond exactly with the metiers/fleets considered in the case study.}
   \item{4}{: If catch at age data is only available at stock level.}
   \item{5}{: If we want to use the data available previously in the \code{FLCatch} objects from \code{flt_obj} to derive catch profiles at age 
            and then apply \code{caaOpt==3} using only one fleet segment, fseg, which represents all the fleets and metiers.
            Note: This approach could lead to a different total catch at age profile derived from the fleets to those in the stocks.}
}}

\item{priceOpt}{A code number to determine the way in wich price at age data are provided.
The option to be used depends on the data availabiltiy, from data rich to data-poor and the following codes are available:
\itemize{ 
   \item{1}{: If price data is available at metier level for all the metiers.}
   \item{2}{: If price data is only available at fleet level.}
   \item{3}{: If price data is disaggregated but the segments do not correspond exactly with the metiers/fleets considered in the case study.}
   \item{4}{: If price data is only available at stock level.}
}}

\item{excel}{Logical. If \code{TRUE} (default), the data in the Excel file is used to create the stucture of the FLFleets object; whereas 
if \code{FALSE}, the \code{flt_obj} object is used instead.}
}
\value{
An \code{FLFleetsExt}.
}
\description{
This function generates an FLFleetsExt object, given the data inputs as arrays. 
             Supported formats are Excel (xls and xlsx) and R format (RData).
}
\seealso{
\code{\link{FLFleetsExt}}, \code{\link{create.biol.arrays}}
}
\author{
Dorleta Garcia & Sonia Sanchez.
}
\keyword{create.fleets.arrays}
