% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_CtrlCreators_fleets.r
\name{create.fleets.ctrl}
\alias{create.fleets.ctrl}
\title{fleets.ctrl object creator}
\usage{
create.fleets.ctrl(flts, n.flts.stks, flts.stksnames, catch.threshold = NULL,
  seasonal.share = NULL, effort.models = NULL, capital.models = NULL,
  catch.models = NULL, price.models = NULL, flq, ...)
}
\arguments{
\item{flts}{character vector with fleet names}

\item{n.flts.stks}{numeric vector with the same length as flts with the  declaration of the number of stocks caugth by each of the fleets.}

\item{flts.stksnames}{character vector with length = sum(n.flts.stks), with the names of the stocks caught by the fleet, the vector must follow the order used in the previous argument. 
\itemize{ 
   \item the  first n.flts.stks[1] elements correspond to the stocks caught by the first fleet in flts
   \item the  following n.flts.stks[2] elements correspond to the stocks caught by the second fleet in flts
   \item and so on.}}

\item{catch.threshold}{if(NULL) => 0.9 for all the stocks (NULL is the default)
else it must be an FLQuant with dim = c(nstks,ny,1,ns,nit)}

\item{effort.models}{characted vector with the same length as flts with the effort model followed by each of the fleet. 
the first element correspond with the effort model of the first fleet in flts, the second with the second and so on.
The default is NULL in which case 'fixedEffort' is used for **all** the fleets.}

\item{capital.models}{characted vector with the same length as flts with the capital model followed by each of the fleet. 
the first element correspond with the capital model of the first fleet in flts, the second with the second and so on.
The default is NULL in which case 'fixedCapital' is used for **all** the fleets.}

\item{catch.models}{characted vector with the same length as sum(n.flts.stks) with the catch model followed by each of the fleet for each stock. 
the first element correspond with the catch model of the first fleet in flts and the first stock in flts.stksnames, the second with the second and so on.
The default is NULL in which case 'CobbDouglasAge' is used for **all** the fleets.}

\item{price.models}{characted vector with the same length as sum(n.flts.stks) with the price model followed by each of the fleet for each stock. 
the first element correspond with the price model of the first fleet in flts and the first stock in flts.stksnames, the second with the second and so on.
The default is NULL in which case 'fixedPrice' is used for **all** the fleets.}

\item{flq}{An flquant to give structure to the FLQuants to be used within the function, 
the dimensioan and dimnames in 'year', 'season' and 'iter' will be used to create the necessary FLQuants.}

\item{...:}{Any extra arguments necessary in the model specific creators. '...' are extracted using 'list(...)', this generates a named list with the extra arguments.
To assure the correct functioning the extra arguments must have a name, for example, elas = FLQuant(1,dimnames = DimsNms).}
}
\value{
A list of lists with the basic structure of the fleets.ctrl object.
}
\description{
It creates the fleets.ctrl object to be used in the call to the main function FLBEIA.
}
