% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_CtrlCreators_Observation.r
\name{create.obs.ctrl}
\alias{create.obs.ctrl}
\title{obs.ctrl object creator}
\usage{
create.obs.ctrl(stksnames, n.stks.inds = NULL, stks.indsnames = NULL,
  stkObs.models = NULL, indObs.models = NULL, immediate = FALSE, ...)
}
\arguments{
\item{stksnames}{A vector with the name of the stocks in the OM.}

\item{n.stks.inds}{numeric vector with the same length as stksnames with the  declaration of the number of abundance indices per stock, the order must be the same used in stksnames.}

\item{stks.indsnames}{NULL or a character vector of length equal to sum(n.stks.inds) with the names of the indices per stock. The order must be the same used in the previous arguments.}

\item{stkObs.models}{A character vector of the same length as stksnames with the name of the model used to observed stock data.}

\item{indObs.models}{A character vector of the same length as stks.indsnames with the name of the model used to generate the abundance indices.}

\item{immediate}{logical, indicating if the warnings should be output immediately.}

\item{...}{any extra arguments necessary in the model specific creators. '...' are extracted using 'list(...)', this generates a named list with the extra arguments.
To assure the correct functioning the extra arguments must have a name.}
}
\value{
A list of lists with the basic structure of the obs.ctrl object.
}
\description{
It creates the obs.ctrl object to be used in the call to the main function FLBEIA.
}
