% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_create.SRs.data.R
\name{create.SRs.data}
\alias{create.SRs.data}
\title{FLBEIA easy conditioning: SRs argument creator}
\usage{
create.SRs.data(yrs, ns, ni, stks.data)
}
\arguments{
\item{yrs}{A vector with c(first.yr,proj.yr, last.yr) where:
\itemize{
    \item first.yr: First year of simulation (number).
    \item proj.yr: First year of projection (number).
    \item last.yr: Last year of projection (number).}}

\item{ns}{Number of seasons (number).}

\item{ni}{Number of iterations (number).}

\item{stks.data}{A list with the name of the stks and the following elements:
\itemize{
    \item stk.unit: Number of units of the stock (number).
    \item  stk.age.min: Minimum age class of the stock (number).
    \item  stk.age.max: Maximum age class of the stock (number).
    \item stk_sr.model: Name of the model to simulate recruitment (character).
    \item stk_params.n: Number of parameters (number).
    \item stk_params.name: Name of the parameters (vector).
    \item stk_params.array:    Parameter values (array).
    \item stk_rec.flq: Recruitment values (FLQuant).
    \item stk_ssb.flq: Spawning stock biomass values (FLQuant).
    \item stk_proportion.flq: Recruitment distribution in each time step as a proportion (FLQuant, values between 0 and 1).
    \item stk_prop.avg.yrs: Historical years to calculate the proportion average (vector).
    \item stk_timelag.matrix: Timelag between the spawning an recruitment (matrix [2, number of seasons]). For details see FLSRsim.
    \item stk_range.plusgroup: Plusgroup age (number).
    \item stk_range.minyear: Minimum year (number).}
Optionals:
\itemize{
    \item stk_uncertainty.flq: Uncertainty (FLQuant).}}
}
\value{
A list of FLSRsim objects.
}
\description{
create.BDs.data function creates a list of FLBDsim objects
}
