% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions_Data_Manipulation.R
\name{extractBRP}
\alias{extractBRP}
\title{Extracts Bpa, Blim, Bmsy, Fpa, Flim and Fmsy reference points from advice.ctrl object (if available),
If not, values are set to NA.}
\usage{
extractBRP(advice.ctrl, stkn)
}
\arguments{
\item{advice.ctrl}{A list with advice controls as the FLBEIA function argument advice.ctrl.}
}
\value{
A new FLBEIA output object with all the iterations joined.
}
\description{
The output can be used for the brp argument of \code{bioSum} function.
}
\seealso{
\code{\link{predictModel}}
}
