% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Summary.R
\name{F_flbeia}
\alias{F_flbeia}
\alias{SSB_flbeia}
\alias{B_flbeia}
\alias{R_flbeia}
\alias{C_flbeia}
\alias{L_flbeia}
\alias{D_flbeia}
\title{Auxiliary summary Functions}
\usage{
F_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

SSB_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

B_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

R_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

C_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

L_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

D_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)
}
\arguments{
\item{FLQuant}{object.}

\item{year.}{}
}
\value{
A FLQuant object.
}
\description{
Auxiliary summary Functions
}
