% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_1_FLBDsim.R
\docType{class}
\name{FLBDsim}
\alias{FLBDsim}
\title{FLBDsim class and methods}
\arguments{
\item{...}{Empty or FLQuants for 'biomass', 'catch' and 'uncertainty' slots and optionally the values for the rest of the slots.}
}
\value{
An object of class FLBDsim.
}
\description{
A class  to simulate the growth of populations aggregated in biomass.
}
\details{
The FLBDsim has the following slots:
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the stock.}

\item{\code{desc}}{A description of the object.}

\item{\code{range}}{The range of the object.}

\item{\code{biomass}}{An FLQuant with to store the biomass of the stock.}

\item{\code{catch}}{An FLQuant with to store the catch of the stock.}

\item{\code{uncertainty}}{An FLQuant with to store the uncertainty that is multiplied to the biomass in every step of the simulation.}

\item{\code{covar}}{An FLQuants with to store the covariates that are part of the growth model.}

\item{\code{params}}{An array with dimension [numb.params, numb.year, numd.season, numb.iteration] with time, year and season and iteration dependent parameters of the growth model.}

\item{\code{alpha}}{A numeric value bigger than one which indicates, in percentage, how big can be the biomass in comparison with the carrying capacity.}

\item{\code{name}}{The name of the object.}

\item{\code{desc}}{Character with the description of the object.}

\item{\code{range}}{A numeric vector with the range of the object as in other FLR objects.}
}}

