% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_Function.R
\name{FLBEIA}
\alias{FLBEIA}
\title{Run the FLBEIA bio-economic simulation model}
\usage{
FLBEIA(biols, SRs = NULL, BDs = NULL, fleets, covars = NULL,
  indices = NULL, advice = NULL, main.ctrl, biols.ctrl, fleets.ctrl,
  covars.ctrl, obs.ctrl, assess.ctrl, advice.ctrl)
}
\arguments{
\item{biols}{An FLBiols object.}

\item{SRs}{A list of FLSRSim objects. One per age structured stock in biols object.}

\item{BDs}{A list of FLSRSim objects. One per biomass dynamic stock in biols object.}

\item{fleets}{An FLFleetsExt object. An extended version of the FLFleet object defined in FLCore.}

\item{covars}{A list of FLQuants used to store any kind of variables that are used within the simulation and are not stored in the standard objects.}

\item{indices}{A list of FLIndices. Each element must correspond with one of the stocks in biols object.}

\item{advice}{A list with two FLQuant elements, TAC and quota.share. TAC is an FLQuant  with quant dimension equal to the number of stocks in biols object, the names used in in the quant dimension must be equal to those used in biols. quota.share is a list with one element per stock in biols object indicating the quota share per stock and fleet. The quant dimension of the elements must be equal to the number of fleets and the names used must be equal to those in fleets objects.}

\item{main.ctrl}{A list with the settings to control the main function (the year range,...).}

\item{biols.ctrl}{A list with the settings to control the biological operating model for each stock (the population dynamic model used,  additional parameters,...)}

\item{fleets.ctrl}{A list with the settings to control the fleets operating model for each fleet (the fleets' short and long term dynamic models used, price model, additional parameters,...)}

\item{covars.ctrl}{A list with the settings to control the covars operating model for each fleet (a dynamic model for each covariable,  additional parameters, ...)}

\item{obs.ctrl}{A list with the settings to control the observation model for each stock (the observation model for the stock, for stock dependent indices, additional parameters, ...)}

\item{assess.ctrl}{A list with the settings to control the specify the assessment model for each stock (the assessment model for the stock and the control parameters used to run the model)}

\item{advice.ctrl}{A list with the settings to control the advice model for each stock (the HCR for each stock, the reference points used in the HCR, additional parameters, ...)}
}
\value{
A list with 8 elements biols, fleets, covars,  advice, stocks, indices,  fleets.ctrl,
}
\description{
Run the FLBEIA bio-economic simulation model
}
