% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R, R/Class_5_FLFleetExt.R
\docType{class}
\name{FLCatchesExt}
\alias{FLCatchesExt}
\alias{vFLCs}
\alias{FLCatchesExt-class}
\alias{FLCatchesExt,ANY-method}
\alias{FLCatchesExt,ANY}
\alias{FLCatchesExt,missing-method}
\alias{FLCatchesExt,missing}
\alias{FLCatchesExt,list-method}
\alias{FLCatchesExt,list}
\alias{is.FLCatchesExt}
\alias{is.FLCatchesExt,ANY-method}
\alias{is.FLCatchesExt,ANY}
\alias{catchNames,FLCatchesExt-method}
\alias{catchNames,FLCatchesExt}
\alias{catchNames,FLMetierExt-method}
\alias{catchNames,FLMetierExt}
\alias{catchNames,FLMetiersExt-method}
\alias{catchNames,FLMetiersExt}
\alias{catchNames,FLFleetExt-method}
\alias{catchNames,FLFleetExt}
\alias{catchNames,FLFleetsExt-method}
\alias{catchNames,FLFleetsExt}
\title{Class FLCatchesExt}
\usage{
vFLCs(object)

\S4method{FLCatchesExt}{ANY}(object, ...)

\S4method{FLCatchesExt}{missing}(object, ...)

\S4method{FLCatchesExt}{list}(object)

is.FLCatchesExt(object, ...)

\S4method{is.FLCatchesExt}{ANY}(object, ...)

\S4method{catchNames}{FLCatchesExt}(object)

\S4method{catchNames}{FLMetierExt}(object)

\S4method{catchNames}{FLMetiersExt}(object)

\S4method{catchNames}{FLFleetExt}(object)

\S4method{catchNames}{FLFleetsExt}(object)
}
\arguments{
\item{object}{An object of class FLCatchExt, list or missing.}

\item{...}{Other objects to be assigned by name to the class slots.}
}
\description{
A list of \code{FLCatchExt} objects.
}
\section{Slots}{
 \describe{
  \item{.Data}{Internal S4 data representation, of class \code{list}.}
  \item{desc}{As textual description of the object contents}
  \item{lock}{Can the object be extended/trimmed? \code{TRUE} or \code{FALSE}.}
  \item{names}{A character vector for the element names} }
}

\seealso{
\code{\link{FLlst}}, \code{\link[base]{list}},
  \code{\link[base]{vector}}, \code{\link{FLCatchExt}}
}
\author{
The FLBEIA Team
}
\keyword{classes}
