% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R, R/Class_3_FLCatchExt.R
\docType{class}
\name{FLCatchesExt-class}
\alias{FLCatchesExt-class}
\alias{FLCatchesExt,ANY-method}
\alias{FLCatchesExt,missing-method}
\alias{FLCatchesExt,list-method}
\alias{is.FLCatchesExt}
\alias{is.FLCatchesExt,ANY-method}
\alias{FLCatchExt-class}
\alias{FLCatchExt,FLQuant-method}
\alias{FLCatchExt,missing-method}
\title{FLcatchExt and FLCatches classes and the methods to construct it.}
\usage{
\S4method{FLCatchesExt}{ANY}(object, ...)

\S4method{FLCatchesExt}{missing}(object, ...)

\S4method{FLCatchesExt}{list}(object)

is.FLCatchesExt(object, ...)

\S4method{is.FLCatchesExt}{ANY}(object, ...)

\S4method{FLCatchExt}{FLQuant}(object, range = "missing", name = "NA",
  desc = character(0), ...)

\S4method{FLCatchExt}{missing}(object, ...)
}
\description{
They extend the FLCatch and FLCatches classes defined in FLFleets package. the FLCatch class includes two extra slots alpha and beta
used in the Cobb-Douglas production functions.
}
\details{
@param  object An object of class FLQuant, missing or FLCatchExt.
 
 @return The constructors return an object of class FLCatchExt.
}
