% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R, R/Class_3_FLCatchExt.R
\docType{class}
\name{FLCatchesExt-class}
\alias{FLCatchesExt-class}
\alias{FLCatchesExt}
\alias{FLCatchesExt,ANY-method}
\alias{FLCatchesExt,}
\alias{ANY}
\alias{FLCatchesExt,missing-method}
\alias{FLCatchesExt,}
\alias{missing}
\alias{FLCatchesExt,list-method}
\alias{FLCatchesExt,}
\alias{list}
\alias{is.FLCatchesExt}
\alias{FLCatchesExt,}
\alias{is.FLCatchesExt,ANY-method}
\alias{FLCatchesExt,}
\alias{is.FLCatchesExt,ANY}
\alias{FLCatchExt}
\alias{FLCatchExt-class}
\alias{FLCatchExt-methods}
\alias{FLQuant,FLCatchExt-method}
\alias{FLCatchExt,FLQuant-method}
\alias{FLCatchExt,missing-method}
\alias{FLCatchExt-missing}
\alias{[,FLCatchExt,ANY,ANY,ANY-method}
\alias{[<-,FLCatchExt,ANY,ANY,FLCatchExt-method}
\alias{[<-,FLCatchExt,ANY,ANY,ANY-method}
\alias{addFLCatch,FLCatchExt,FLCatchExt-method}
\alias{addFLCatch,}
\alias{FLCatchExt-methods,}
\alias{FLCatchExt,}
\alias{FLCatchExt}
\alias{catchNames,FLCatchExt-method}
\alias{catchNames,}
\alias{FLCatchExt}
\alias{catchNames<-,FLCatchExt,character-method}
\alias{catchNames<-,}
\alias{FLCatchExt,}
\alias{character}
\title{FLcatchExt and FLCatchesExt classes and the methods to construct it.}
\usage{
\S4method{FLCatchesExt}{ANY}(object, ...)

\S4method{FLCatchesExt}{missing}(object, ...)

\S4method{FLCatchesExt}{list}(object)

is.FLCatchesExt(object, ...)

\S4method{is.FLCatchesExt}{ANY}(object, ...)

\S4method{FLCatchExt}{FLQuant}(object, range = "missing", name = "NA",
  desc = character(0), ...)

\S4method{FLCatchExt}{missing}(object, ...)

\S4method{[}{FLCatchExt,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[}{FLCatchExt,ANY,ANY,FLCatchExt}(x, i, j, k, l, m, n, ...) <- value

\S4method{addFLCatch}{FLCatchExt,FLCatchExt}(e1, e2)

\S4method{catchNames}{FLCatchExt}(object)

\S4method{catchNames}{FLCatchExt,character}(object) <- value
}
\arguments{
\item{object, x}{An object of class FLQuant, missing or FLCatchExt.}

\item{...}{Other objects to be assigned by name to the class slots}

\item{range}{Numerical vector with min, max, plusgroup, minyear and maxyear elements as in FLStock object.}

\item{name}{The name of the stock.}

\item{desc}{The description of the object.}

\item{i, j, k, l, m, n}{subindices}

\item{drop}{logical. Should the dimesions be dropped?}
}
\value{
The constructors return an object of class FLCatchExt.
}
\description{
They extend the FLCatch and FLCatches classes defined in FLFleets package. the FLCatch class includes two extra slots alpha and beta
used in the Cobb-Douglas production functions.
}
\details{
The FLCatchExt object contains a representation of the catch of a  fish stock as constructed for the purposes of fleet dynamic modelling. 
   This includes information on removals (i.e. landings and discards), selectivity, weight, price and catch production parameters (catchability and elasticities).
}
