% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_5_FLFleetExt.R
\docType{class}
\name{FLFleetExt}
\alias{FLFleetExt}
\alias{FLFleetExt-class}
\alias{FLFleetExt-methods}
\alias{FLFleetExt,FLFleetExt-method}
\alias{FLFleetExt,FLMetiersExt-method}
\alias{FLFleetExt,FLMetierExt-method}
\alias{FLFleetExt,FLCatchesExt-method}
\alias{FLFleetExt,FLCatchExt-method}
\alias{FLFleetExt,missing-method}
\alias{FLFleetExt,FLFleetExt-missing}
\alias{[,FLFleetExt,ANY,missing,ANY-method}
\alias{[,FLFleetExt,ANY,missing-method}
\alias{[,FLFleetExt,ANY,ANY,ANY-method}
\alias{[,FLFleetExt,ANY,ANY-method}
\alias{[[,FLFleetExt,ANY,missing-method}
\title{FLFleetExt class and the methods to construct it.}
\usage{
\S4method{FLFleetExt}{FLMetiersExt}(object, ...)

\S4method{FLFleetExt}{FLMetierExt}(object, ...)

\S4method{FLFleetExt}{FLCatchesExt}(object, ...)

\S4method{FLFleetExt}{FLCatchExt}(object, ...)

\S4method{FLFleetExt}{FLFleetExt}(object, metier, catch, ...)

\S4method{FLFleetExt}{missing}(object, ...)

\S4method{[}{FLFleetExt,ANY,missing,ANY}(x, i, drop = FALSE)

\S4method{[}{FLFleetExt,ANY,ANY,ANY}(x, i, j, drop = FALSE)

\S4method{[[}{FLFleetExt,ANY,missing}(x, i, drop = FALSE)
}
\arguments{
\item{object, x}{An object of class FLQuant, missing, FLFleetExt, 
FLCatchExt, FLCatchesExt or FLMetierExt.}

\item{...}{Other objects to be assigned by name to the class slots}

\item{metier}{A name of one of the elements in FLMetiersExt object.}

\item{catch}{A name of one of the elements in FLCatchesExt object.}

\item{i, j}{subindices.}

\item{drop}{If TRUE, deletes the dimensions of an array which have only one level.}
}
\value{
The constructors return an object of class FLFleetExt.
}
\description{
It extends the FLFleetExt class defined in FLFleet package. 
The only difference is that that the metiers slot is a FLMetiersExt object.
}
\details{
The FLFleetExt object contains a representation of a fishing fleet as constructed for the purposes of fleet dynamic modelling. 
   This includes information on effort, fixed-cost, capacity, crew-share, metiers and variable costs.
}
\section{Slots}{

\describe{
\item{\code{effort}}{An FLQuant with the effort of the fleet. The effort can have any units (e.g. number of fishing days, trips, hooks,...)}

\item{\code{fcost}}{An FLQuant with the fixed costs of the fleet. 
These costs should be given by vessel and the number of vessels by fleet must be included in the covars object.}

\item{\code{capacity}}{An FLQuant with the capacity of the fleet. Same units as in slot effort must be used.}

\item{\code{crewshare}}{An FLQuant with the crewshare of the fleet. Where crewshare is the percentage of revenues that goes to the crew.}

\item{\code{metiers}}{A FLMetiersExt with information on the fleet's metiers.}

\item{\code{name}}{The name of the stock.}

\item{\code{desc}}{A description of the object.}

\item{\code{range}}{The range as in other FLR objects: c("min","max","plusgroup","minyear","maxyear").}
}}

