% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R
\docType{class}
\name{FLFleetsExt}
\alias{FLFleetsExt}
\alias{vFLFs}
\alias{FLFleetsExt-class}
\alias{FLFleetsExt,ANY-method}
\alias{FLFleetsExt,ANY}
\alias{FLFleetsExt,missing-method}
\alias{FLFleetsExt,FLFleetExt-missing}
\alias{FLFleetsExt,list-method}
\alias{FLFleetsExt,list}
\alias{is.FLFleetsExt}
\alias{is.FLFleetsExt,ANY-method}
\alias{is.FLFleetsExt,ANY}
\title{Class FLFleetsExt}
\usage{
vFLFs(object)

\S4method{FLFleetsExt}{ANY}(object, ...)

\S4method{FLFleetsExt}{missing}(object, ...)

\S4method{FLFleetsExt}{list}(object)

is.FLFleetsExt(object, ...)

\S4method{is.FLFleetsExt}{ANY}(object, ...)
}
\arguments{
\item{object}{An object of class FLFleetExt, list or missing.}

\item{...}{Other objects to be assigned by name to the class slots.}
}
\description{
A list of \code{FLFleetExt} objects.
}
\section{Slots}{
 \describe{
  \item{.Data}{Internal S4 data representation, of class \code{list}.}
  \item{desc}{As textual description of the object contents}
  \item{lock}{Can the object be extended/trimmed? \code{TRUE} or \code{FALSE}.}
  \item{names}{A character vector for the element names} }
}

\seealso{
\code{\link{FLlst}}, \code{\link[base]{list}},
  \code{\link[base]{vector}}, \code{\link{FLFleetExt}}
}
\author{
The FLBEIA Team
}
\keyword{classes}
