% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_4_FLMetierExt.R
\docType{class}
\name{FLMetierExt}
\alias{FLMetierExt}
\alias{FLMetierExt-class}
\alias{FLMetierExt-methods}
\alias{FLMetierExt,FLMetierExt-method}
\alias{FLMetiersExt,FLMetiersExt-methods}
\alias{FLMetierExt,FLCatchExt-method}
\alias{FLMetierExt,FLCatchesExt-method}
\alias{FLMetierExt,FLQuant-method}
\alias{FLMetierExt,missing-method}
\alias{[,FLMetierExt,ANY,missing,ANY-method}
\alias{[,FLMetierExt,ANY,missing-method}
\alias{[[,FLMetierExt,ANY,missing-method}
\title{FLMetierExt class and the methods to construct it.}
\usage{
\S4method{FLMetierExt}{FLCatchExt}(catches, gear = "NA", ...)

\S4method{FLMetierExt}{FLCatchesExt}(catches, gear = "NA", ...)

\S4method{FLMetierExt}{FLQuant}(catches, gear = "NA", ...)

\S4method{FLMetierExt}{missing}(catches, gear = "NA", ...)

\S4method{[}{FLMetierExt,ANY,missing,ANY}(x, i, drop = FALSE)

\S4method{[[}{FLMetierExt,ANY,missing}(x, i, drop = FALSE)
}
\arguments{
\item{catches, x}{An object of class FLQuant, missing or FLCatchExt.}

\item{gear}{A character vector with the name of the gear used in the metier.}

\item{...}{Other objects to be assigned by name to the class slots}

\item{i}{subindices.}

\item{drop}{If TRUE, deletes the dimensions of an array which have only one level.}
}
\value{
The constructors return an object of class FLMetierExt.
}
\description{
It extends the FLMetier class defined in FLFleet package. 
The only difference is that that the catches slot is a FLCatchesExt object.
}
\details{
The FLMetierExt object contains a representation of the metier of a fishing fleet as constructed for the purposes of fleet dynamic modelling. 
   This includes information on effortshare and variable costs.
}
\section{Slots}{

\describe{
\item{\code{gear}}{A character with the gear name of a fleet's metier.}

\item{\code{effshare}}{An FLQuant with the effort share of a fleet's metier relative to fleet's total effort 
(the sum of all metiers effshares for a fleet must sum 1).}

\item{\code{vcost}}{An FLQuant with the varible costs of a fleet's metier.
These costs should be given by vessel and based on the effort units used for the fleet's effort (within the FLFleetExt object).
The number of vessels by fleet must be included in the covars object.}

\item{\code{catches}}{A FLCatchesExt with information on the fleet's metier catches.}

\item{\code{name}}{The name of the stock.}

\item{\code{desc}}{A description of the object.}

\item{\code{range}}{The range as in other FLR objects: c("min","max","plusgroup","minyear","maxyear").}
}}

