% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R
\docType{class}
\name{FLMetiersExt}
\alias{FLMetiersExt}
\alias{vFLMs}
\alias{FLMetiersExt-class}
\alias{FLMetiersExt,ANY-method}
\alias{FLMetiersExt,ANY}
\alias{FLMetiersExt,missing-method}
\alias{FLMetiersExt,missing}
\alias{FLMetiersExt,list-method}
\alias{FLMetiersExt,list}
\alias{is.FLMetiersExt}
\alias{FLMetiersExt,is.FLMetiersExt}
\alias{is.FLMetiersExt,ANY-method}
\alias{FLMetiersExt,is.FLMetiersExt,ANY}
\title{Class FLMetiersExt}
\usage{
vFLMs(object)

\S4method{FLMetiersExt}{ANY}(object, ...)

\S4method{FLMetiersExt}{missing}(object, ...)

\S4method{FLMetiersExt}{list}(object)

is.FLMetiersExt(object, ...)

\S4method{is.FLMetiersExt}{ANY}(object, ...)
}
\arguments{
\item{object}{An object of class FLMetierExt, list or missing.}

\item{...}{Other objects to be assigned by name to the class slots.}
}
\description{
A list of \code{FLMetierExt} objects.
}
\section{Slots}{
 \describe{
  \item{.Data}{Internal S4 data representation, of class \code{list}.}
  \item{desc}{As textual description of the object contents}
  \item{lock}{Can the object be extended/trimmed? \code{TRUE} or \code{FALSE}.}
  \item{names}{A character vector for the element names} }
}

\seealso{
\code{\link{FLlst}}, \code{\link[base]{list}},
  \code{\link[base]{vector}}, \code{\link{FLMetierExt}}
}
\author{
The FLBEIA Team
}
\keyword{classes}
