% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_JoinIterations.R
\name{joinIter}
\alias{joinIter}
\title{Joins the iterations of several FLBEIA outputs}
\usage{
joinIter(
  objnam,
  files,
  directory = NULL,
  Niters = 1,
  elements = "all",
  advice.ext = "TAC",
  fleets.ctrl.ext = "seasonal.share"
)
}
\arguments{
\item{objnam}{Character. The name of the object that will be joined.
The object must be the output of FLBEIA function.}

\item{files}{A character vector with the names of the files from which \code{objnam} will be taken.}

\item{directory}{The directory were the files are stored. Default value is the current directory.}

\item{Niters}{A numeric vector with the number of iterations per object.
If length=1, then it is assumed that all objects have the same number of iterations.}

\item{elements}{The elements of the objects that must be joined. 
The default is to join all the objects.}

\item{advice.ext}{Character. The element from \code{advice} object that will be replaced. Default is 'TAC'.}

\item{fleets.ctrl.ext}{Character. The element from \code{fleets.ctrl} object that will be replaced. Default is 'seasonal.share'.}
}
\value{
A new FLBEIA output object with all the iterations joined.
}
\description{
joinIter function allows to join different outputs of FBEIA simulations. 
For example when simulations are run iteration by iteration separately, 
with this function we can merged all the iterations into one FLBEIA output object.
}
\note{
The files must contain a single object (named as \code{objnam} value).
}
