\name{FLBEIA}
\alias{FLBEIA}
\alias{FLBEIA-package}

\alias{advice.mp}
\alias{age2ageDat}
\alias{age2agePop}
\alias{age2bioDat}
\alias{age2bioPop}
\alias{age2unit}
\alias{ageInd}
\alias{aneHCRE}
\alias{aneRec_pil}
\alias{annualTAC}
\alias{annexIVHCR}
\alias{ASPG}
\alias{assessment.mp}
\alias{B_flbeia}
\alias{BDPG}
\alias{BDsim}
\alias{bio2bioDat}
\alias{bio2bioPop}
\alias{bioInd}
\alias{biols.om}
\alias{biomass.tg}
\alias{bioSum}
\alias{C_flbeia}
\alias{capacityRest.SMFB}
\alias{catch.tg}
\alias{catchFlSum }
\alias{catchMtSum}
\alias{catchStock}
\alias{catchStock.f}
\alias{catchWStock}
\alias{catchWStock.f}
\alias{checkBDTarget}
\alias{checkDims}
\alias{chkBDTargetQuantity}
\alias{chkBDTrgtArrayIters}
\alias{chkFwdBDControl}
\alias{CFPMSYHCR}
\alias{cobbDoug}
\alias{cobbDougInv}
\alias{CobbDouglasFit}
\alias{CobbDouglasBio}
\alias{CobbDouglasAge}
\alias{CobbDouglasBio.effort}
\alias{CobbDouglasAge.effort}
\alias{CobbDouglasBio.CAA}
\alias{CobbDouglasAge.CAA}
\alias{covars.om}
\alias{CorrectCatch}
\alias{costs_flbeia}
\alias{createFleetExtAccesors}
\alias{create.advice.ctrl}
\alias{create.assess.ctrl}
\alias{create.biols.ctrl}
\alias{create.fleets.ctrl}
\alias{create.covars.ctrl }
\alias{create.obs.ctrl}
\alias{create.advice.data}
\alias{create.BDs.data}
\alias{create.biols.data}
\alias{create.fleets.data}
\alias{create.F2CatchHCR.ctrl}
\alias{create.indices.data}
\alias{calculate.CDparam}
\alias{create.list.stks.flq}
\alias{create.list.stks.flqa}
\alias{create.SRs.data}
\alias{ctRec_alb}
\alias{D_flbeia}
\alias{discards.tg}
\alias{discnStock}
\alias{discStock}
\alias{discStock.f}
\alias{discWStock}
\alias{discWStock.f}
\alias{ecoSum}
\alias{effortMtSum}
\alias{effRule.SMFB}
\alias{effRule.SSFB}
\alias{elasticPrice}
\alias{equal.flq.Dimnames }
\alias{exponential}
\alias{F2CatchHCR}
\alias{f.discards.tg}
\alias{f.landings.tg}
\alias{f.tg}
\alias{F_flbeia}
\alias{fillIterFLF}
\alias{findF}
\alias{fixedAdvice}
\alias{fixedCapital}
\alias{fixedCovar}
\alias{fixedEffort}
\alias{fixedPopulation}
\alias{fixedPrice}
\alias{FLBDsim}
\alias{FLBEIA}
\alias{fleets.om}
\alias{FLSRsim}
\alias{FLXSAnew}
\alias{FroeseHCR}
\alias{fwdBD}
\alias{fwdBDControl}
\alias{getPlural}
\alias{getQuantity}
\alias{ghlHCR}
\alias{gmeanHistGrowth}
\alias{hockstick}
\alias{initPop_lifeHist}
\alias{IcesHCR}
\alias{joinIter}
\alias{L_flbeia}
\alias{landings.tg}
\alias{landStock}
\alias{landStock.f}
\alias{landWStock}
\alias{landWStock.f}
\alias{MAPHCR}
\alias{MaxProfit}
\alias{matrixBDTarget}
\alias{neaMAC_ltmp}
\alias{NoAssessment}
\alias{NoObsIndex}
\alias{NoObsStock}
\alias{Obs.stk.nage}
\alias{Obs.stk.bio}
\alias{Obs.disc.nage}
\alias{Obs.fec}
\alias{Obs.FSolver}
\alias{Obs.land.nage}
\alias{Obs.stk.wgt}
\alias{Obs.disc.bio}
\alias{Obs.land.bio}
\alias{Obs.disc.wgt}
\alias{Obs.land.wgt}
\alias{Obs.nmort}
\alias{observation.mp}
\alias{PellaTom}
\alias{perfectObs}
\alias{pilRec_ane}
\alias{plot.catchFl}
\alias{plot.eco}
\alias{plot.biols}
\alias{plot.fleets}
\alias{pope}
\alias{price_flbeia}
\alias{propagateFLF}
\alias{quantityNms}
\alias{R_flbeia}
\alias{redfishRecModel}
\alias{remQ.SSFB}
\alias{replaceRec}
\alias{revenue_flbeia}
\alias{SCD}
\alias{seasonshare.CAA}
\alias{SelAtAge}
\alias{showArray}
\alias{SMFB}
\alias{SRsim}
\alias{SSB_flbeia}
\alias{ssb.get}
\alias{SSFB}
\alias{stock.fleetInfo}
\alias{summary_flbeia}
\alias{tdiscStock}
\alias{tlandStock}
\alias{total.catch.stock.df}
\alias{total.landings.stock.df}
\alias{total.discards.stock.df}
\alias{totfcost_flbeia }
\alias{totvcost_flbeia}
\alias{trgtNms}
\alias{unit2age}
\alias{updateCatch}
\alias{updateQS.SMFB}
\alias{validFwdBDControl}
\alias{vFLBDSRsims}
\alias{vFLCs}
\alias{vFLFs}
\alias{vFLMs}
\alias{wtadStock}
\alias{wtalStock}
\alias{XSAboot}
\alias{xtabs2}

\alias{addFLCatch}
\alias{addFLCatch-methods}
\alias{addFLCatch,FLCatchExt,FLCatchExt-method}

\alias{catchNames}
\alias{catchNames-methods}
\alias{catchNames,FLCatchExt-method}
\alias{catchNames,FLCatchesExt-method}
\alias{catchNames,FLFleetExt-method}
\alias{catchNames,FLFleetsExt-method}
\alias{catchNames,FLMetierExt-method}
\alias{catchNames,FLMetiersExt-method}


\alias{catchNames<-}
\alias{catchNames<--methods}
\alias{catchNames<-,FLCatchExt,character-method}

\alias{FLCatchesExt}
\alias{FLCatchesExt-methods}
\alias{FLCatchesExt,ANY-method}
\alias{FLCatchesExt,list-method}
\alias{FLCatchesExt,missing-method}

\alias{FLCatchExt}
\alias{FLCatchExt-methods}
\alias{FLCatchExt,missing-method}
\alias{FLCatchExt,FLQuant-method}

\alias{FLFleetExt}
\alias{FLFleetExt-methods}
\alias{FLFleetExt,FLCatchExt-method}
\alias{FLFleetExt,FLCatchesExt-method}
\alias{FLFleetExt,FLFleetExt-method}
\alias{FLFleetExt,FLMetierExt-method}
\alias{FLFleetExt,FLMetiersExt-method}
\alias{FLFleetExt,missing-method}


\alias{FLFleetsExt}
\alias{FLFleetsExt-methods}
\alias{FLFleetsExt,ANY-method}
\alias{FLFleetsExt,missing-method}
\alias{FLFleetsExt,list-method}

\alias{FLMetierExt}
\alias{FLMetierExt-methods}
\alias{FLMetierExt,FLCatchExt-method}
\alias{FLMetierExt,FLCatchesExt-method}
\alias{FLMetierExt,FLQuant-method}
\alias{FLMetierExt,missing-method}

\alias{FLMetiersExt}
\alias{FLMetiersExt-methods}
\alias{FLMetiersExt,ANY-method}
\alias{FLMetiersExt,missing-method}
\alias{FLMetiersExt,list-method}

\alias{is.FLBDSRsims}
\alias{is.FLBDSRsims-methods}
\alias{is.FLBDSRsims,ANY-method}

\alias{is.FLCatchesExt}
\alias{is.FLCatchesExt-methods}
\alias{is.FLCatchesExt,ANY-method}

\alias{is.FLFleetsExt}
\alias{is.FLFleetsExt-methods}
\alias{is.FLFleetsExt,ANY-method}

\alias{is.FLMetiersExt}
\alias{is.FLMetiersExt-methods}
\alias{is.FLMetiersExt,ANY-method}

\alias{[,FLCatchExt,ANY,ANY,ANY-method}
\alias{[,FLFleetExt,ANY,ANY,ANY-method}
\alias{[,FLFleetExt,ANY,missing,ANY-method}
\alias{[,FLMetierExt,ANY,missing,ANY-method}
\alias{[<-,FLCatch,ANY,ANY,FLCatch-method}
\alias{[[,FLFleetExt,ANY,missing-method}
\alias{[[,FLMetierExt,ANY,missing-method}

\alias{FLBDsim-class}
\alias{FLSRsim-class}
\alias{FLCatchesExt-class}
\alias{FLCatchExt-class}
\alias{FLFleetExt-class}
\alias{FLFleetsExt-class}
\alias{FLMetierExt-class}
\alias{FLMetiersExt-class}

\title{FLBEIA}
\description{FLBEIA
}
\seealso{
  \code{\link{FLCore}}.
}
\keyword{package}
