% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brps_seasonal.R
\name{plotBRPsson}
\alias{plotBRPsson}
\title{YPR plots for a seasonal model}
\usage{
plotBRPsson(obj, pdfnm = "Fbar_vs_SPR.pdf")
}
\arguments{
\item{obj}{The output of brpsson function or a data.frame with the same structure as the runs object in the 
output list of brpsson function.}

\item{pdfnm}{The name of the pdf document where plots are going to be saved.}
}
\value{
A pdf for each stock with plots.
}
\description{
Function for generating YPR plots for reference points when having a seasonal model
}
\seealso{
\code{\link{brpsson}}
}
