% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCatchFl.R
\name{plotCatchFl}
\alias{plotCatchFl}
\title{Plots with fleets data}
\usage{
plotCatchFl(fleets, advice, pdfnm)
}
\arguments{
\item{fleets}{A FLFleets object.}

\item{advice}{A list with two FLQuant elements, TAC and quota.share.}

\item{pdfnm}{The name for the pdf document will be 'Catch-' and pdfnm.}
}
\value{
One pdf for with plots (landings, discards and price by fleet).
}
\description{
Return a pdf with plots using FLFleets and advice objects.
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(one)
plotCatchFl(fleets = oneFl,advice = oneAdv,pdfnm = "s0")
}

}
