% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEco.R
\name{plotEco}
\alias{plotEco}
\title{Plots with fleets data}
\usage{
plotEco(fleets, pdfnm)
}
\arguments{
\item{fleets}{A FLFleets object.}

\item{pdfnm}{The name for the pdf document will be "Eco" and pdfnm separated by a line.}
}
\value{
A pdf with capacity, costs, effort, profits by fleet.
}
\description{
Return a pdf with plots using FLFleets object.
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(one)
plotEco(s0$fleets,'s0')
}
}
