% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFLBiols.R
\name{plotFLBiols}
\alias{plotFLBiols}
\title{Plots with biological data}
\usage{
plotFLBiols(biols, pdfnm)
}
\arguments{
\item{biols:}{a FLBiols object}

\item{pdfnm:}{the name for the pdf document will be stock's name and pdfnm separated by a line.}
}
\value{
A pdf for each stock with plots.
}
\description{
Return a pdf for each stock with plots using FLBiols object.
}
\details{
\itemize{
     \item{Each pdf contains biomass in numbers at age, mean weight at age, fecundity, natural mortality, maturity, spawning, recruitment and spawning stock biomass} 
}
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(one)
s0 <- FLBEIA(biols = oneBio,       # FLBiols object with one FLBiol element for stk1.
               SRs = oneSR,        # A list with one FLSRSim object for stk1.
               BDs = NULL,         # No Biomass Dynamic populations in this case.
            fleets = oneFl,        # FLFleets object with on fleet.
            covars = NULL,         # covars not used
           indices = NULL,         # indices not used 
            advice = oneAdv,       # A list with two elements 'TAC' and 'quota.share'
         main.ctrl = oneMainC,     # A list with one element to define the start and end of the simulation.
        biols.ctrl = oneBioC,      # A list with one element to select the model to simulate the stock dynamics.
       fleets.ctrl = oneFlC,       # A list with several elements to select fleet dynamic models and store additional parameters.
       covars.ctrl = NULL,         # covars control not used 
          obs.ctrl = oneObsC,      # A list with one element to define how the stock observed ("PerfectObs").
       assess.ctrl = oneAssC,      # A list with one element to define how the stock assessment model used ("NoAssessment").
       advice.ctrl = oneAdvC) 
plotFLBiols(s0$biols, 's0')
}
}
