% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFLFleets.R
\name{plotFLFleets}
\alias{plotFLFleets}
\title{Plots with fleets data}
\usage{
plotFLFleets(fleets, pdfnm)
}
\arguments{
\item{fleets}{A FLFleets object.}

\item{pdfnm}{The name for the pdf document will be the fleet's name and pdfnm separated by a line.}
}
\value{
A pdf for each fleet with plots.
}
\description{
For each fleet, return a pdf with plots using FLFleets object.
}
\details{
For each fleet, the pdf contains plots of:
\itemize{
     \item Catch, discards, landings, capacity, crewshare, 
     effort, fcost, effshare
     \item For each metier: landings and discards at age in numbers 
     and mean weight, alpha, beta and catch.q 
}
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(one)
s0 <- FLBEIA(biols = oneBio,       # FLBiols object with one FLBiol element for stk1.
               SRs = oneSR,        # A list with one FLSRSim object for stk1.
               BDs = NULL,         # No Biomass Dynamic populations in this case.
            fleets = oneFl,        # FLFleets object with on fleet.
            covars = NULL,         # covars not used
           indices = NULL,         # indices not used 
            advice = oneAdv,       # A list with two elements 'TAC' and 'quota.share'
         main.ctrl = oneMainC,     # A list with one element to define the start and end of the simulation.
        biols.ctrl = oneBioC,      # A list with one element to select the model to simulate the stock dynamics.
       fleets.ctrl = oneFlC,       # A list with several elements to select fleet dynamic models and store additional parameters.
       covars.ctrl = NULL,         # covars control not used 
          obs.ctrl = oneObsC,      # A list with one element to define how the stock observed ("PerfectObs").
       assess.ctrl = oneAssC,      # A list with one element to define how the stock assessment model used ("NoAssessment").
       advice.ctrl = oneAdvC) 
plotFLFleets(s0$fleets,'s0')
}
}

