% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotfltStkSum.R
\name{plotfltStkSum}
\alias{plotfltStkSum}
\title{Plots with fltStkSum data}
\usage{
plotfltStkSum(obj, pdfnm)
}
\arguments{
\item{obj}{FLBEIA output}

\item{pdfnm}{The name for the pdf document will be 'fltStkSum-' and pdfnm.}
}
\value{
One pdf with plots on landings, discards,catch,price,quotaUpt,tacshare,discRat,quota by fleet and stock.
}
\description{
Return a pdf with plots with the outputs fltStkSum, usint the output of FLBEIA.
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(res_flbeia)
plotfltStkSum(obj=oneRes,pdfnm = "oneRes")
}

}
