% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Summary.R
\name{revenue_flbeia}
\alias{revenue_flbeia}
\alias{costs_flbeia}
\alias{totvcost_flbeia}
\alias{totfcost_flbeia}
\alias{taxcost_flbeia}
\alias{price_flbeia}
\title{Economic summary functions.}
\usage{
revenue_flbeia(fleet, fleet.ctrl, advice)

costs_flbeia(fleet, advice, covars, fleet.ctrl, flnm = NULL)

totvcost_flbeia(fleet, advice, covars, fleet.ctrl)

totfcost_flbeia(fleet, covars, flnm = NULL)

taxcost_flbeia(fleet, fleet.ctrl, advice)

price_flbeia(fleet, stock)
}
\arguments{
\item{fleet}{An element of a FLFleetsExt object (FLFleetExt).}

\item{fleet.ctrl}{List with tax-related controls (requires fleet.ctrl$taxes = TRUE/FALSE and tax function with corresponding parameters by stock).}

\item{advice}{List of two FLQuants, with TAC and quota share for all the stocks.}

\item{covars}{List of FLQuants with information on covariates.}

\item{flnm}{Name of the fleet.}

\item{stock}{An FLStock object.}
}
\description{
These functions provide summary results of costs, prices and revenues. Provided data can be disaggregated by fleet or by metier depending on the selected function.
}
\details{
\itemize{
      \item{revenue_flbeia}{ computes the revenue by fleet and metier. The revenue is computed as
       landings (weight) multiplied by the price, minus taxes.}
      \item{taxcost_flbeia}{ computes taxes (to be deducted from the revenue).}
      \item{costs_flbeia}{ computes total costs as the sum of fixed and variable costs.}
      \item{totvcost_flbeia}{ computes the variable costs including crew share costs .}
      \item{totfcost_flbeia}{ computes the total costs by vessel.}
      \item{price_flbeia}{ computes the price by stock.} 
     }
}
