% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_1a1_Stock_Recruitment_functions_segregmix.R
\name{segremix}
\alias{segremix}
\alias{segregmix}
\title{Stock-Recruitment models in FLBEIA: segrexmix - Mixture of 2 segmented regression stock-recruitment models fit}
\usage{
segregmix()
}
\description{
Model formulation:
}
\details{
\deqn{R = \mathbf{ifelse}(S \leq b, a S, \mathbf{ifelse}(S \leq B, a b, A B uncAdd))}
      { R = ifelse(S <= b, a*S, ifelse(S <= B, a*b, A*B*uncAdd))}
\emph{a} is the slope of the recruitment for stock levels at or below \emph{b},
\eqn{a b}{a*b} is the mean recruitment for stock levels above \emph{b} and at or below \emph{B}, and
\eqn{A B uncAdd}{A*B*uncAdd} is the mean recruitment for stock levels above \emph{B}.
Where \emph{a, b, A, B} > 0.

Additional stock-recruitment (SR) models to the ones provided in FLCore package.
}
\seealso{
\link{SRModels}, \linkS4class{FLSR}, \linkS4class{FLModel}
}
\author{
The FLBEIA Team
}
