% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions_remove_units.R
\name{setUnitsNA}
\alias{setUnitsNA}
\title{Function to remove the units from the FLQuants of FLBiol or FLFleetExt objects}
\usage{
setUnitsNA(obj)
}
\arguments{
\item{obj}{FLBiol or FLFleetExt object}
}
\value{
The same object with the units equal to NA in all the FLQuant slots.
}
\description{
Function to remove the units from the FLQuants of FLBiol or FLFleetExt objects
}
\examples{
\dontrun{

data(one)

fl <- setUnitsNA(oneBio)
fl <- setUnitsNA(oneFl)
}
}
