% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FLFleets.R
\name{stock.fleetInfo}
\alias{stock.fleetInfo}
\title{stock.fleetInfo.}
\usage{
stock.fleetInfo(fleets)
}
\arguments{
\item{fleets}{is an objetc of class FLFleetsExt.}
}
\description{
\code{stock.fleetInfo} Return a matrix with rownames equal to the 
name of the stocks and colnames equal to fleet names and metier names.
}
\details{
Return a matrix with rownames equal to the stocks names and colnames equal to names of fleet and metier.
If element (i,j) is equal to 0, the stock (i) is not caught by fleet/metier (j).
}
\examples{
\dontrun{
data(multi)
stock.fleetInfo(fl1)
}
}
