% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Summary.R, R/ecoSum_DAMARA_function.R
\name{summary_flbeia}
\alias{summary_flbeia}
\alias{bioSum}
\alias{ecoSum}
\alias{catchFlSum}
\alias{catchMtSum}
\alias{effortMtSum}
\alias{ecoSum_damara}
\title{Summary of the FLBEIA output}
\usage{
summary_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

bioSum(obj, years = dimnames(obj$biols[[1]]@n)$year)

ecoSum(fleets, flnms = "all", years, covars = NULL)

catchFlSum(fleets, advice, flnms = "all", stknms = "all", years)

catchMtSum(fleets, flnms = "all", stknms = "all", years)

effortMtSum(fleets, flnms, years)

ecoSum_damara(fleets, flnms = "all", years, covars = NULL)
}
\arguments{
\item{obj}{The output of the FLBEIA function.}

\item{years}{The period of years.}

\item{fleets}{An FLFleetsExt object. An extended version of the FLFleet object defined in FLCore.}

\item{flnms}{Names of the fleets.}

\item{covars}{A list of FLQuants used to store any kind of variables that are used within the simulation and are not stored in the standard objects.}

\item{advice}{A list with two FLQuant elements, TAC and quota.share. TAC is an FLQuant  with quant dimension equal to the number of stocks in biols object, the names used in in the quant dimension must be equal to those used in biols. quota.share is a list with one element per stock in biols object indicating the quota share per stock and fleet. The quant dimension of the elements must be equal to the number of fleets and the names used must be equal to those in fleets objects.}

\item{stknms}{Names of the stocks.}
}
\value{
A data frame.
}
\description{
Summarize the results of the simulation in data frames.
}
\details{
\itemize{
     \item{Summary_flbeia:} The data frame contains five columns: stock, years, iteration, 
     indicator and value. The indicators are: recruitment, ssb, f, biomass, catch, landings and discards.
     \item{bioSum:} The data frame contains biological results. The columns are stocks, 
     years, it, indicators and value. 
     \item{ecoSum:} The colums of the data frame are: year, quarter, stock, fleet, iter, profits, 
     capacity, costs, discards, effort and landings.
     \item{ecoSum_damara:} ecoSum for Damara project.
     \item{effortMtSum:} The colums of the data frame are: year, quarter, fleet, metier, iter, 
     effort and effshare.
     \item{catchFlSum:} The colums of the data frame are: year, quarter, stock, fleet, iter, 
     landings, discards, price and tacshare.
     \item{catchMtSum:} The colums of the data frame are: year, quarter, stock, fleet, metier, iter, 
     landings, discards and price.
     
}
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(one)
s0 <- FLBEIA(biols = oneBio,       # FLBiols object with one FLBiol element for stk1.
               SRs = oneSR,        # A list with one FLSRSim object for stk1.
               BDs = NULL,         # No Biomass Dynamic populations in this case.
            fleets = oneFl,        # FLFleets object with on fleet.
            covars = NULL,         # covars not used
           indices = NULL,         # indices not used 
            advice = oneAdv,       # A list with two elements 'TAC' and 'quota.share'
         main.ctrl = oneMainC,     # A list with one element to define the start and end of the simulation.
        biols.ctrl = oneBioC,      # A list with one element to select the model to simulate the stock dynamics.
       fleets.ctrl = oneFlC,       # A list with several elements to select fleet dynamic models and store additional parameters.
       covars.ctrl = NULL,         # covars control not used 
          obs.ctrl = oneObsC,      # A list with one element to define how the stock observed ("PerfectObs").
       assess.ctrl = oneAssC,      # A list with one element to define how the stock assessment model used ("NoAssessment").
       advice.ctrl = oneAdvC) 
summary_flbeia(obj=s0)
}
}
