% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions_Data_Manipulation.R
\name{unit2age}
\alias{unit2age}
\title{Translates unit dimension in an FQuant into age dimension}
\usage{
unit2age(Q)
}
\arguments{
\item{Q}{A FLQuant.
The object must be the output of FLBEIA function.}
}
\value{
A 4-dimensional array with length c(d[1]*d[3], d[2], d[4], d[6]), 
        where d is the dimension of the FLQuant.
}
\description{
This function transforms a FQuant with several 'unit' dimension into 
an array an unique 'unit' dimension and no 'area' dimension. Moving 'unit' to 'age' dimension.
This is usefull when we use the 'unit' dimension to store the different seasonal cohorts.
}
\note{
The files must contain a single object (named as \code{objnam} value).
}
