% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions_Data_Manipulation.R
\name{updateFLBiols}
\alias{updateFLBiols}
\title{Update the FLBiols object to the FLCore 2.6}
\usage{
updateFLBiols(biols)
}
\arguments{
\item{biols}{A FLBiols object.}
}
\value{
A new FLBEIA output object with all the iterations joined.
}
\description{
Updates an old FLBiols (where slots fec and mat from each FLBiol are FLQuants), into the new version of the class,
where slots fec and mat are of class \code{predictModel}.
}
\seealso{
\code{\link{predictModel}}
}
