% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions_xtabs2.R
\name{xtabs2}
\alias{xtabs2}
\title{Cross tabulation}
\usage{
xtabs2(
  formula = ~.,
  data = parent.frame(),
  subset,
  sparse = FALSE,
  na.action,
  exclude = c(NA, NaN),
  drop.unused.levels = FALSE
)
}
\arguments{
\item{formula}{a \link[stats]{formula} object with the cross-classifying variables
    (separated by \code{+}) on the right hand side (or an object which
    can be coerced to a formula).  Interactions are not allowed.  On the
    left hand side, one may optionally give a vector or a matrix of
    counts; in the latter case, the columns are interpreted as
    corresponding to the levels of a variable.  This is useful if the
    data have already been tabulated, see the examples below.}

\item{data}{an optional matrix or data frame (or similar: see
    \code{\link[stats]{model.frame}}) containing the variables in the
    formula \code{formula}.  By default the variables are taken from
    \code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations
    to be used.}

\item{sparse}{logical specifying if the result should be a
    \emph{sparse} matrix, i.e., inheriting from
    \code{\link[Matrix:sparseMatrix-class]{sparseMatrix}}%\linkS4class{sparseMatrix}.
    Only works for two factors (since there
    are no higher-order sparse array classes yet).
  }

\item{na.action}{a function which indicates what should happen when
    the data contain \code{\link{NA}}s.  If unspecified, and
    \code{addNA} is true, this is set to \code{\link[stats]{na.pass}}.  When it
    is \code{na.pass} and \code{formula} has a left hand side (with
    counts), \code{\link{sum}(*, na.rm = TRUE)} is used instead of
    \code{sum(*)} for the counts.}

\item{exclude}{a vector of values to be excluded when forming the
    set of levels of the classifying factors.}

\item{drop.unused.levels}{a logical indicating whether to drop unused
    levels in the classifying factors.  If this is \code{FALSE} and
    there are unused levels, the table will contain zero marginals, and
    a subsequent chi-squared test for independence of the factors will
    not work.}
}
\value{
An array.
}
\description{
Create a contingency table as an array, similar to \code{xtabs}.
However, whereas \code{xtabs} converts NA-s in 0, \code{xtabs2} does not touch NA values.
Moreover, while output class in \code{xtabs} is c("xtabs", "table"), 
it is an array in \code{xtabs2}.
}
