% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Summary.R
\name{bioSum}
\alias{bioSum}
\title{Summary of the FLBEIA output}
\usage{
bioSum(obj, stocks = "all", years = dimnames(obj$biols[[1]]@n)$year,
  long = TRUE, scenario = "bc")
}
\arguments{
\item{obj}{The output of the FLBEIA function.}

\item{years}{The period of years.}

\item{long}{The data frame should be constructed using long or wide format?}

\item{flnms}{Names of the fleets.}

\item{stknms}{Names of the stocks.}
}
\value{
The data frames can be of wide or long format. In long format all the indicators are in the same column.
 There is one column, indicator, for the name of the indicator and a second one value for the numerica value of the indicator.
 In the wide format each of the indicators correspond with one column in the data frame. 
 The long format it is recommendable to work with ggplot2 functions for example and the wide format 
 it is more efficient for memory allocation and speed of computations.

A data frame.
}
\description{
Summarize the results of the simulation in data frames.
}
\details{
\itemize{
     \item{bioSum:} Data frame with the biological indicators The columns are stocks, 
     years, it, indicators and value. 
     \item{fleetSum:} The colums of the data frame are: year, quarter, stock, fleet, iter, profits, 
     capacity, costs, discards, effort and landings.
     \item{ecoSum_damara:} ecoSum for Damara project.
     \item{metierSum:} The colums of the data frame are: year, quarter, fleet, metier, iter, 
     effort and effshare.
     \item{stockFleetSum:} The colums of the data frame are: year, quarter, stock, fleet, iter, 
     landings, discards, price and tacshare.
     \item{stockMetierSum:} The colums of the data frame are: year, quarter, stock, fleet, metier, iter, 
     landings, discards and price.
     
}
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(one)
s0 <- FLBEIA(biols = oneBio,       # FLBiols object with one FLBiol element for stk1.
               SRs = oneSR,        # A list with one FLSRSim object for stk1.
               BDs = NULL,         # No Biomass Dynamic populations in this case.
            fleets = oneFl,        # FLFleets object with on fleet.
            covars = NULL,         # covars not used
           indices = NULL,         # indices not used 
            advice = oneAdv,       # A list with two elements 'TAC' and 'quota.share'
         main.ctrl = oneMainC,     # A list with one element to define the start and end of the simulation.
        biols.ctrl = oneBioC,      # A list with one element to select the model to simulate the stock dynamics.
       fleets.ctrl = oneFlC,       # A list with several elements to select fleet dynamic models and store additional parameters.
       covars.ctrl = NULL,         # covars control not used 
          obs.ctrl = oneObsC,      # A list with one element to define how the stock observed ("PerfectObs").
       assess.ctrl = oneAssC,      # A list with one element to define how the stock assessment model used ("NoAssessment").
       advice.ctrl = oneAdvC) 
       
s0_bio <- bioSum(obj=s0)
s0_eco <- ecoSum(obj=s0)

}
}
