% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brps_seasonal.R
\name{brpsson}
\alias{brpsson}
\title{Reference points for a seasonal model}
\usage{
brpsson(
  stk,
  B0,
  R0,
  rec.ss = 1,
  ssb.ss = 1,
  sr_model,
  sr_params,
  Fprop = rep(1/dim(stk)[4], dim(stk)[4]),
  Fscan = seq(0, 4, by = 0.01),
  oldest = 100,
  nrun = 200,
  tol = 0.01
)
}
\arguments{
\item{stk}{An FLStock object.}

\item{B0}{The value of the virgin biomass.}

\item{R0}{The expected recrutiment in the virgin population.}

\item{rec.ss}{The recruitment season (numeric). Default value = 1.}

\item{ssb.ss}{The spawning season (numeric). Default value = 1.}

\item{sr_model}{A character with the name of the model to simulate the recruitment process.}

\item{sr_params}{A named vector with the SR parameter values.}

\item{Fprop}{A vector with the same length as the number of seasons with the proportion of F by season. 
By default the same proportion in all the seasons is assumed.}

\item{Fscan}{A vector with the F values to be simulated. 
By default, will scan values between 0 and 4 (by 0.01 increments).}

\item{oldest}{The maximum age class to be considered. Default 100.}

\item{nrun}{The maximum number of years to project forward until equilibrium.}

\item{tol}{The desired accuracy.}
}
\value{

}
\description{
Function to estimate reference points when having a seasonal model
}
\examples{

library(FLBEIA)
data(multistk) # object with 2 seasons and 3 iterations

stk <- trim( multistk, year=1, iter = 1) 

# loop for different catch proportions by season
for (p in seq(0.1,0.9,0.1)) {
  fruns <- brpsson( stk, B0=1e+05, R0=27489766, rec.ss=2, ssb.ss=2, 
                    sr_model="bevholt", sr_params=c( a = 29988835.109, b = 9090.909), 
                    Fprop = c(p, 1-p))
} 

plotBRPsson( fruns, pdfnm="stk_Fbar_vs_SPR.pdf")
 
}
\seealso{
\code{\link{plotBRPsson}}
}
