% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_create.BDs.data.R
\name{create.BDs.data}
\alias{create.BDs.data}
\title{FLBEIA easy conditioning: BDs argument creator}
\usage{
create.BDs.data(yrs, ns, ni, stks.data)
}
\arguments{
\item{yrs}{A vector with c(first.yr,proj.yr, last.yr) where:
\itemize{
    \item first.yr: First year of simulation (number).
    \item proj.yr: First year of projection (number).
    \item last.yr: Last year of projection (number).}}

\item{ns}{Number of seasons (number).}

\item{ni}{Number of iterations (number).}

\item{stks.data}{A list with the name of the stks and the following elements:
\itemize{
    \item stk.unit: Number of units of the stock (number).
    \item stk.age.min: Minimum age class of the stock (number).
    \item stk.age.max: Maximum age class of the stock (number).
    \item stk_bd.model: Name of the model to simulate biomass dinamics of the stock (character).
    \item stk_params.name: Name of the parameters (vector).
    \item stk_params.array:    Parameter values (array).
    \item stk_biomass.flq: Biomass values (FLQuant).
    \item stk_catch.flq: Catch values (FLQuant).
    \item stk_range.plusgroup: Plusgroup age (numeric).
    \item stk_range.minyear: Minimum year (numeric).
    \item stk_alpha: Maximum variability of carrying capacity.}
Optionals:
\itemize{
    \item stk_gB.flq: Surplus production (FLQuant).
    \item stk_uncertainty.flq: Uncertainty (FLQuant).}}
}
\value{
A list of FLBDsim objects.
}
\description{
create.BDs.data function creates a list of FLBDsim objects.
}
