% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_create.biols.data.R
\name{create.biols.data}
\alias{create.biols.data}
\title{FLBEIA easy conditioning: biols argument creator}
\usage{
create.biols.data(yrs, ns, ni, stks.data)
}
\arguments{
\item{yrs}{A vector with c(first.yr,proj.yr, last.yr) where
\itemize{
     \item first.yr: First year of simulation (number).
     \item proj.yr:  First year of projection (number).
     \item last.yr:  Last year of projection (number).}}

\item{ns}{Number of seasons (number).}

\item{ni}{Number of iterations (number).}

\item{stks.data}{A list with the name of the stks and the following elements:
\itemize{
     \item  stk.unit: Number of units of the stock (number). 
     \item  stk.age.min: Minimum age class of the stock (number).
     \item  stk.age.max: Maximum age class of the stock (number).
     \item  stk_n.flq: Numbers at age in the population(FLQuant).
     \item  stk_wt.flq: Weight at age of an individual (FLQuant).
     \item  stk_m.flq: Mortality rate at age of the population (FLQuant).    
     \item  stk_fec.flq: Fecundity at age    (FLQuant).    
     \item  stk_mat.flq: Percentage of mature individuals at age    (FLQuant).    
     \item  stk_spwn.flq: Proportion of time step at which spawning ocurrs    (FLQuant).
     \item  stk.range.plusgroup: Plusgroup age (number).
     \item  stk.range.minyear: Minimum year (number).
     \item  stk.range.maxyear: Maximum year (number).
     \item  stk_range.minfbar: Minimum age to calculate average fishing mortality (number).
     \item  stk_range.maxfbar: Maximum age to calculate average fishing mortality (number).
     \item  stk_biol.proj.avg.yrs:    Historic years to calculate the average of spwn, fec, m and wt for the projection (vector).}}
}
\value{
An FLBiol object
}
\description{
create.biols.data function creates an FLBiols object.
}
