% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditioning_create.fleets.data.R
\name{create.fleets.data}
\alias{create.fleets.data}
\title{FLBEIA easy conditioning: fleets argument creator}
\usage{
create.fleets.data(yrs, ns, ni, fls.data, stks.data)
}
\arguments{
\item{yrs}{A vector with c(first.yr,proj.yr, last.yr) where
\itemize{
     \item first.yr: First year of simulation (number).
     \item proj.yr: First year of projection (number).
     \item last.yr: Last year of projection (number).}}

\item{ns}{Number of seasons (number).}

\item{ni}{Number of iterations (number).}

\item{fls.data}{A list with the name of the fleets and the following elements:
\itemize{
     \item fl.met: Name of the metiers in the fleet 'fl' (vector).
     \item fl.met.stks: Name of the stocks in the metier 'met' and fleet 'fl' (vector).
     \item fl_effort.flq: 'fl' fleet's effort (FLQuant).
     \item fl.met_effshare.flq: 'fl' fleet and 'met' metier's effort share (FLQuant).
     \item fl.met.stk_landings.n.flq: 'fl' fleet,'met' metier and 'stk' stock's landings in numbers at age.
     \item fl_proj.avg.yrs: Historic years to calculate the average of effort, fcost, crewshare, capacity in 'fl' fleet (vector) in the projection years.}
Optionals:
\itemize{
     \item fl_capacity.flq: 'fl' fleet's capacity (FLQuant).
     \item fl_fcost.flq: 'fl' fleet's fixed cost (FLQuant).
     \item fl_crewshare.flq: 'fl' fleet's crewshare (FLQuant).
     \item fl.met_vcost.flq: 'fl' fleet and 'met' metier's variable costs (FLQuant).
     \item fl.met.stk_landings.wt.flq: 'fl' fleet,'met' metier and 'stk' stock's mean weight of landings at age.
     \item fl.met.stk_discards.n.flq: 'fl' fleet,'met' metier and 'stk' stock's discards in numbers at age (FLQuant).
     \item fl.met.stk_discards.wt.flq: 'fl' fleet,'met' metier and 'stk' stock's mean weight of discards at age.
     \item fl.met.stk_price.flq: 'fl' fleet,'met' metier and 'stk' stock's price at age (FLQuant).
     \item fl.met.stk_alpha.flq:    'fl' fleet,'met' metier and 'stk' stock's Cobb Douglass alpha parameter (FLQuant).
     \item fl.met.stk_beta.flq: 'fl' fleet,'met' metier and 'stk' stock's Cobb Douglass beta parameter (FLQuant).
     \item fl.met.stk_catch.q.flq: 'fl' fleet,'met' metier and 'stk' stock's Cobb Douglass catch.q parameter (FLQuant).
     \item fl.met_proj.avg.yrs: Historic years to calculate the average of effshare,vcost for 'fl' fleet and 'met' metier in the projection period (vector).
     \item fl.met.stk_proj.avg.yrs: Historic years to calculate the average of landings.wt, discards.wt, landings.sel, discards.sel 
                                  alpha,beta,catch.q for 'fl' fleet, 'met' metier and 'stk' stock in the projection years (vector).}}

\item{stks.data}{A list with the name of the stocks and with the next elements:
\itemize{
     \item  stk.unit: Number of units of the stock (number). 
     \item  stk.age.min: Minimum age of the stock (number).
     \item  stk.age.max: Maximum age of the stock (number).}
Optionals:
\itemize{
     \item  stk_wt.flq: Mean weight at age of an individual (FLQuant). Required if fl.met.stk_landings.wt.flq is not defined.
     \item  stk_n.flq: Numbers at age in the population(FLQuant). Required if Cobb Douglas parameters are not defined.
     \item  stk_gB.flq: Biomass growth for the stock modeled in biomass (FLQuant). Required if Cobb Douglas parameters are not defined.}}
}
\value{
An FLFleetsExt object.
}
\description{
create.fleets.data function creates an FLFleetsExt object
}
