% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Summary.R
\name{F_flbeia}
\alias{F_flbeia}
\alias{SSB_flbeia}
\alias{B_flbeia}
\alias{R_flbeia}
\alias{C_flbeia}
\alias{L_flbeia}
\alias{D_flbeia}
\alias{summary_flbeia}
\title{Biological summary functions}
\usage{
F_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

SSB_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

B_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

R_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

C_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

L_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

D_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

summary_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)
}
\arguments{
\item{obj}{The output of the FLBEIA function.}

\item{years}{The  years for which the indicators are extracted.}
}
\value{
B_flbeia, F_flbeia... return an array with three dimensions (stock, year and iter).
The summary_flbeia function returns an array with 4 dimensions (stock, year, iter, indicator) with the value of all the indicators.
}
\description{
These functions return the biomass (B), fishing mortality (F),  spawning stock biomass (SSB), recruitment (R), catches (C), landings (L) and discards (D) indicators.
}
\details{
\itemize{
      \item{B_flbeia}{ this function computes SSB.}
      \item{F_flbeia}{ ithis function computes fishing mortiality.}
      \item{SSB_flbeia}{ this function computes spawning stock biomass by species.}
      \item{R_flbeia}{ this function computes recruitment by stock. If the stock is defined by age this function the recruiment is computed. ;
                       If the stock is follows a biomass dynamics, this function gives the growth.}
      \item{C_flbeia}{ this function computes catches by fleets and stock.} 
      \item{L_flbeia}{ this function computes landings by fleets and stock.}
      \item{D_flbeia}{ ithis function computes the discards by fleets and stock.}
     }
}
