% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_1_FLBDsim.R
\docType{class}
\name{FLBDsim}
\alias{FLBDsim}
\title{FLBDsim class and methods}
\arguments{
\item{...}{Empty or FLQuants for 'biomass', 'catch' and 'uncertainty' slots and optionally the values for the rest of the slots.}
}
\value{
An object of class FLBDsim.
}
\description{
A class  to simulate the growth of populations aggregated in biomass.
}
\details{
The FLBDsim has the following slots:
\itemize{
   \item{name}{ The name of the stock.}
   \item{desc}{ A description of the object.}
   \item{range}{ The range of the object.}
   \item{biomass}{ An FLQuant with to store the biomass of the stock.} 
   \item{catch}{ An FLQuant with to store the catch of the stock.} 
   \item{uncertainty}{ An FLQuant with to store the uncertainty that is multiplied to the biomass in every step of the simulation.} 
   \item{biomass}{ An FLQuant with to store the biomass of the object.} 
   \item{biomass}{ An FLQuant with to store the biomass of the object.} 

}
}
