% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R, R/Class_3_FLCatchExt.R
\docType{methods}
\name{FLCatchesExt,ANY-method}
\alias{FLCatchesExt,ANY-method}
\alias{FLCatchExt-class}
\title{FLcatchExt and FLCatches classes and the methods to construct it.}
\usage{
\S4method{FLCatchesExt}{ANY}(object, ...)
}
\description{
They extend the FLCatch and FLCatches classes defined in FLFleets package. the FLCatch class includes two extra slots alpha and beta
used in the Cobb-Douglas production functions.
}
\details{
@param  object An object of class FLQuant, missing or FLCatchExt.
 
 @return The constructors return an object of class FLCatchExt.
}
