% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_3_FLCatchExt.R
\docType{class}
\name{FLCatchExt}
\alias{FLCatchExt}
\alias{FLCatchExt-class}
\alias{FLCatchExt-methods}
\alias{FLQuant,FLCatchExt-method}
\alias{FLCatchExt,FLQuant-method}
\alias{FLCatchExt,missing-method}
\alias{FLCatchExt-missing}
\alias{[,FLCatchExt,ANY,ANY,ANY-method}
\alias{[<-,FLCatchExt,ANY,ANY,FLCatchExt-method}
\alias{[<-,FLCatchExt,ANY,ANY,ANY-method}
\alias{catchNames<-}
\alias{addFLCatch,FLCatchExt,FLCatchExt-method}
\alias{addFLCatch,FLCatchExt,FLCatchExt}
\alias{catchNames,FLCatchExt-method}
\alias{catchNames,FLCatchExt}
\alias{catchNames<-,FLCatchExt,character-method}
\alias{catchNames<-,FLCatchExt,character}
\title{FLCatchExt class and the methods to construct it.}
\usage{
\S4method{FLCatchExt}{FLQuant}(object, range = "missing", name = "NA", desc = character(0), ...)

\S4method{FLCatchExt}{missing}(object, ...)

\S4method{[}{FLCatchExt,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[}{FLCatchExt,ANY,ANY,FLCatchExt}(x, i, j, k, l, m, n, ...) <- value

catchNames(object, ...) <- value

\S4method{addFLCatch}{FLCatchExt,FLCatchExt}(e1, e2)

\S4method{catchNames}{FLCatchExt}(object)

\S4method{catchNames}{FLCatchExt,character}(object) <- value
}
\arguments{
\item{object, x}{An object of class FLQuant, missing or FLCatchExt.}

\item{range}{Numerical vector with min, max, plusgroup, minyear and maxyear elements as in FLStock object.}

\item{name}{The name of the stock.}

\item{desc}{The description of the object.}

\item{...}{Other objects to be assigned by name to the class slots.}

\item{i, j, k, l, m, n}{subindices}

\item{drop}{logical. Should the dimesions be dropped?}

\item{value}{Value or values to be assigned to the particular FLQuant or FLCatchExt slot.}

\item{e1, e2}{FLCatchExt objects, where e2 is incorporated into e1 (see addFLCatch).}
}
\value{
The constructors return an object of class FLCatchExt.
}
\description{
It extends the FLCatch class defined in FLFleet package. 
   The FLCatch class includes two extra slots alpha and beta used in the Cobb-Douglas production functions.
}
\details{
The FLCatchExt object contains a representation of the catch of a fish stock as constructed for the purposes of fleet dynamic modelling. 
   This includes information on removals (i.e. landings and discards), selectivity, weight, price and catch production parameters (catchability and elasticities).
}
\section{Slots}{

\describe{
\item{\code{landings}}{An FLQuant with the total landings in weight of the stock.}

\item{\code{landings.n}}{An FLQuant with the landings in numbers at age of the stock.}

\item{\code{landings.wt}}{An FLQuant with the weight at age of the landings.}

\item{\code{discards}}{An FLQuant with the total discards in weight of the stock.}

\item{\code{discards.n}}{An FLQuant with the discards in numbers at age of the stock.}

\item{\code{discards.wt}}{An FLQuant with the weight at age of the discards.}

\item{\code{landings.sel,discards.sel}}{An FLQuant with the landing/discard ogive of the metier for this stock 
(i.e. landings.sel corresponds to the proportion of catches landed).
Elements must be between 0 and 1, with discards.sel = 1-landings.sel.}

\item{\code{catch.q}}{An FLQuant with the catchability at age of the stock for the corresponding metier. 
This is the catchability used in the catch production model.}

\item{\code{price}}{An FLQuant with the price at age of the stock.}

\item{\code{alpha}}{An FLQuant with the elasticity parameter at age of the stock for the corresponding metier. 
This is one of the parameters used in the catch production model.}

\item{\code{beta}}{An FLQuant with the elasticity parameter at age of the stock for the corresponding metier. 
This is one of the parameters used in the catch production model.}

\item{\code{name}}{The name of the stock.}

\item{\code{desc}}{A description of the object.}

\item{\code{range}}{The range as in other FLR objects: c("min","max","plusgroup","minyear","maxyear").}
}}

