% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R, R/Class_3_FLCatchExt.R
\docType{class}
\name{FLCatchesExt-class}
\alias{FLCatchesExt-class}
\alias{FLCatchesExt}
\alias{FLCatchesExt,ANY-method}
\alias{FLCatchesExt,}
\alias{ANY}
\alias{FLCatchesExt,missing-method}
\alias{FLCatchesExt,}
\alias{missing}
\alias{FLCatchesExt,list-method}
\alias{FLCatchesExt,}
\alias{list}
\alias{is.FLCatchesExt}
\alias{FLCatchesExt,}
\alias{is.FLCatchesExt,ANY-method}
\alias{FLCatchesExt,}
\alias{is.FLCatchesExt,ANY}
\alias{FLCatchExt}
\alias{FLCatchExt-class}
\alias{FLCatchExt-methods}
\alias{FLQuant,FLCatchExt-method}
\alias{FLCatchExt,FLQuant-method}
\alias{FLCatchExt,missing-method}
\alias{FLCatchExt-missing}
\alias{[,FLCatchExt,ANY,ANY,ANY-method}
\alias{[<-,FLCatchExt,ANY,ANY,FLCatchExt-method}
\alias{[<-,FLCatchExt,ANY,ANY,ANY-method}
\alias{addFLCatch,FLCatchExt,FLCatchExt-method}
\alias{addFLCatch,}
\alias{FLCatchExt-methods,}
\alias{FLCatchExt,}
\alias{FLCatchExt}
\alias{catchNames,FLCatchExt-method}
\alias{catchNames,}
\alias{FLCatchExt}
\alias{catchNames<-,FLCatchExt,character-method}
\alias{catchNames<-,}
\alias{FLCatchExt,}
\alias{character}
\title{FLCatchExt and FLCatchesExt classes and the methods to construct it.}
\usage{
\S4method{FLCatchesExt}{ANY}(object, ...)

\S4method{FLCatchesExt}{missing}(object, ...)

\S4method{FLCatchesExt}{list}(object)

is.FLCatchesExt(object, ...)

\S4method{is.FLCatchesExt}{ANY}(object, ...)

\S4method{FLCatchExt}{FLQuant}(object, range = "missing", name = "NA",
  desc = character(0), ...)

\S4method{FLCatchExt}{missing}(object, ...)

\S4method{[}{FLCatchExt,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[}{FLCatchExt,ANY,ANY,FLCatchExt}(x, i, j, k, l, m, n, ...) <- value

\S4method{addFLCatch}{FLCatchExt,FLCatchExt}(e1, e2)

\S4method{catchNames}{FLCatchExt}(object)

\S4method{catchNames}{FLCatchExt,character}(object) <- value
}
\arguments{
\item{object, x}{An object of class FLQuant, missing or FLCatchExt.}

\item{...}{Other objects to be assigned by name to the class slots}

\item{range}{Numerical vector with min, max, plusgroup, minyear and maxyear elements as in FLStock object.}

\item{name}{The name of the stock.}

\item{desc}{The description of the object.}

\item{i, j, k, l, m, n}{subindices}

\item{drop}{logical. Should the dimesions be dropped?}
}
\value{
The constructors return an object of class FLCatchExt.
}
\description{
They extend the FLCatch and FLCatches classes defined in FLFleets package. the FLCatch class includes two extra slots alpha and beta
used in the Cobb-Douglas production functions.
}
\details{
The FLCatchExt object contains a representation of the catch of a  fish stock as constructed for the purposes of fleet dynamic modelling. 
   This includes information on removals (i.e. landings and discards), selectivity, weight, price and catch production parameters (catchability and elasticities).
}
\section{Slots}{

\describe{
\item{\code{landings}}{An FLQuant of dimension [1,num. years, num. units, numb.season, 1, numb. iters] with the total landings in weight of the stock.}

\item{\code{landings.n}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the  landings in numbers at age of the stock.}

\item{\code{landings.wt}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the weight at age of the landings.}

\item{\code{landings.sel}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the retention ogive of the metier for this stock. Elements must be between 0 and 1. landings.sel = 1-discards.sel.}

\item{\code{discards}}{An FLQuant of dimension [1,num. years, num. units, numb.season, 1, numb. iters] with the total discards in weight of the stock.}

\item{\code{discards.n}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the discards in numbers at age of the stock.}

\item{\code{discards.wt}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the weight at age of the discards}

\item{\code{discards.sel}}{landings.sel An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the discard ogive of the metier for this stock. Elements must be between 0 and 1. discards.sel. = 1-landings.sel.}

\item{\code{catch.q}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the catchability at age of the stock for the corresponding metier. This is the catchability used in the catch production model.}

\item{\code{price}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the price at age of the stock.}

\item{\code{alpha}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the elsaticity parameter at age of the stock for the corresponding metier. This is one of the parameters used in the catch production model.}

\item{\code{beta}}{An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the elsaticity parameter at age of the stock for the corresponding metier. This is one of the parameters used in the catch production model}

\item{\code{name}}{The name of the stock.}

\item{\code{desc}}{A description of the object.}

\item{\code{range}}{The range as in other FLR objects. c("min","max","plusgroup","minyear","maxyear")}
}}

