% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R, R/Class_5_FLFleetExt.R
\docType{class}
\name{FLFleetsExt-class}
\alias{FLFleetsExt-class}
\alias{FLFleetsExt}
\alias{FLFleetsExt,ANY-method}
\alias{FLFleetsExt,ANY}
\alias{FLFleetsExt,missing-method}
\alias{FLFleetsExt,FLFleetExt-missing}
\alias{FLFleetsExt,list-method}
\alias{FLFleetsExt,list}
\alias{is.FLFleetsExt}
\alias{is.FLFleetsExt,ANY-method}
\alias{is.FLFleetsExt,ANY}
\alias{FLFleetExt}
\alias{FLFleetExt-class}
\alias{FLFleetExt-methods}
\alias{FLFleetExt,FLFleetExt-method}
\alias{FLFleetExt,FLMetiersExt-method}
\alias{FLFleetExt,FLMetierExt-method}
\alias{FLFleetExt,FLCatchesExt-method}
\alias{FLFleetExt,FLCatchExt-method}
\alias{FLFleetExt,FLFleetExt-method}
\alias{FLFleetExt,missing-method}
\alias{FLFleetExt,FLFleetExt-missing}
\alias{[,FLFleetExt,ANY,missing,ANY-method}
\alias{[,FLFleetExt,ANY,missing-method}
\alias{[,FLFleetExt,ANY,ANY,ANY-method}
\alias{[,FLFleetExt,ANY,ANY-method}
\alias{[[,FLFleetExt,ANY,missing-method}
\title{FLFleetExt and FLFleetsExt classes and the methods to construct it.}
\usage{
\S4method{FLFleetsExt}{ANY}(object, ...)

\S4method{FLFleetsExt}{missing}(object, ...)

\S4method{FLFleetsExt}{list}(object)

is.FLFleetsExt(object, ...)

\S4method{is.FLFleetsExt}{ANY}(object, ...)

\S4method{FLFleetExt}{FLMetiersExt}(object, ...)

\S4method{FLFleetExt}{FLMetierExt}(object, ...)

\S4method{FLFleetExt}{FLCatchesExt}(object, ...)

\S4method{FLFleetExt}{FLCatchExt}(object, ...)

\S4method{FLFleetExt}{FLFleetExt}(object, metier, catch, ...)

\S4method{FLFleetExt}{missing}(object, ...)

\S4method{[}{FLFleetExt,ANY,missing,ANY}(x, i, drop = FALSE)

\S4method{[}{FLFleetExt,ANY,ANY,ANY}(x, i, j, drop = FALSE)

\S4method{[[}{FLFleetExt,ANY,missing}(x, i, drop = FALSE)
}
\arguments{
\item{object}{An object of class FLQuant, missing or FLMetierExt.}

\item{...}{Other objects to be assigned by name to the class slots}

\item{metier}{A name of one of the elements in FLMetiers object.}

\item{catch}{A name of one of the elements in FLCatches object.}

\item{range}{Numerical vector with min, max, plusgroup, minyear and maxyear elements as in FLStock object.}

\item{name}{The name of the fleet.}

\item{desc}{The description of the object.}

\item{metiers}{An object of class FLMetierExt or FLMetiersExt.}
}
\value{
The constructors return an object of class FLFleetExt.
}
\description{
They extend the FLFleetExt and FLFleetsExt classes defined in FLFleets package. The only different is that 
that the metiers slot is a FLMetiersExt object
}
\details{
The FLFleetExt object contains a representation of a  fishing fleet as constructed for the purposes of fleet dynamic modelling. 
   This includes information on effort, fixed-cost, capacity,  crew-share, metiers and variable costs.
}
