% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_0a_FLlst-beia.R, R/Class_4_FLMetierExt.R
\docType{class}
\name{FLMetiersExt-class}
\alias{FLMetiersExt-class}
\alias{FLMetiersExt}
\alias{FLMetiersExt,ANY-method}
\alias{FLMetiersExt,ANY}
\alias{FLMetiersExt,missing-method}
\alias{FLMetiersExt,missing}
\alias{FLMetiersExt,list-method}
\alias{FLMetiersExt,list}
\alias{is.FLMetiersExt}
\alias{FLMetiersExt,is.FLMetiersExt}
\alias{is.FLMetiersExt,ANY-method}
\alias{FLMetiersExt,is.FLMetiersExt,ANY}
\alias{FLMetierExt}
\alias{FLMetierExt-class}
\alias{FLMetierExt-methods}
\alias{FLMetierExt,FLMetierExt-method,}
\alias{FLMetiersExt,}
\alias{FLMEtiersExt-methods}
\alias{FLMetierExt,FLCatchExt-method}
\alias{FLMetierExt,FLCatchesExt-method}
\alias{FLMetierExt,FLQuant-method}
\alias{FLMetierExt,missing-method}
\alias{[,FLMetierExt,ANY,missing,ANY-method}
\alias{[,FLMetierExt,ANY,missing-method}
\alias{[[,FLMetierExt,ANY,missing-method}
\title{FLMetierExt and FLMetiersExt classes and the methods to construct it.}
\usage{
\S4method{FLMetiersExt}{ANY}(object, ...)

\S4method{FLMetiersExt}{missing}(object, ...)

\S4method{FLMetiersExt}{list}(object)

is.FLMetiersExt(object, ...)

\S4method{is.FLMetiersExt}{ANY}(object, ...)

\S4method{FLMetierExt}{FLCatchExt}(catches, gear = "NA", ...)

\S4method{FLMetierExt}{FLCatchesExt}(catches, gear = "NA", ...)

\S4method{FLMetierExt}{FLQuant}(catches, gear = "NA", ...)

\S4method{FLMetierExt}{missing}(catches, gear = "NA", ...)

\S4method{[}{FLMetierExt,ANY,missing,ANY}(x, i, drop = FALSE)

\S4method{[[}{FLMetierExt,ANY,missing}(x, i, drop = FALSE)
}
\arguments{
\item{object}{An object of class FLQuant, missing or FLCatchExt.}

\item{...}{Other objects to be assigned by name to the class slots}

\item{catches}{An object of class FLCatchesExt.}

\item{gear}{A character vector with the name of the gear used in the metier.}

\item{range}{Numerical vector with min, max, plusgroup, minyear and maxyear elements as in FLStock object.}

\item{name}{The name of the metier.}

\item{desc}{The description of the object.}
}
\value{
The constructors return an object of class FLMetierExt.
}
\description{
They extend the FLMetier and FLMetiers classes defined in FLFleets package. The only different is that 
that the catches slot is a FLCatchesExt object
}
\details{
The FLMetierExt object contains a representation of the metier of a  fishing fleet as constructed for the purposes of fleet dynamic modelling. 
   This includes information on effortshare and variable costs.
}
