% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions_Observation_Model.R,
%   R/Functions_FLFleet.R, R/Functions_FLFleets.R
\name{tlandStock}
\alias{tlandStock}
\alias{tdiscStock}
\alias{landStock.f}
\alias{discStock.f}
\alias{catchStock.f}
\alias{landWStock.f}
\alias{discWStock.f}
\alias{catchWStock.f}
\alias{landStock}
\alias{discStock}
\alias{catchStock}
\alias{landWStock}
\alias{discWStock}
\alias{catchWStock}
\title{Extract landings, discard or catch at age}
\usage{
tlandStock(obj, stknm)

tdiscStock(obj, stknm)

landStock.f(obj, stock)

discStock.f(obj, stock)

catchStock.f(obj, stock)

landWStock.f(obj, stock)

discWStock.f(obj, stock)

catchWStock.f(obj, stock)

landStock(obj, stock)

discStock(obj, stock)

catchStock(obj, stock)

landWStock(obj, stock)

discWStock(obj, stock)

catchWStock(obj, stock)
}
\arguments{
\item{obj}{a FLFleetExt (.f) or FLFleetsExt object}

\item{stknm}{stock name}

\item{stock}{stock name}
}
\value{
A FLQuant object with landings, discards or catch (total or at age).
}
\description{
Extract landings, discards or catch at age of a stock from a FLFleetExt or FLFleetsExt object.
}
\details{
\itemize{
     \item{landStock.f:} Extract total landings at age of a stock from a FLFleetExt obj.
     \item{discStock.f:} Extract total discards at age of a stock from a FLFleetExt obj. 
     \item{catchStock.f:} Extract total catch at age of a stock from a FLFleetExt obj.    
     \item{landWStock.f:} Extract total landings at age in weight of a stock from a FLFleetExt obj.
     \item{discWStock.f:} Extract total discards at age in weight of a stock from a FLFleetExt obj.
     \item{catchWStock.f:} Extract total catch at age in weight of a stock from a FLFleetExt obj. 
     \item{landStock:} Extract total landings at age of a stock from a FLFleetExts obj.
     \item{discStock:} Extract total discards at age of a stock from a FLFleetExts obj. 
     \item{catchStock:} Extract total catch at age of a stock from a FLFleetExts obj.    
     \item{landWStock:} Extract total landings at age in weight of a stock from a FLFleetExts obj.
     \item{discWStock:} Extract total discards at age in weight of a stock from a FLFleetExts obj.
     \item{catchWStock:} Extract total catch at age in weight of a stock from a FLFleetExts obj.                       
     \item{tlandStock:} Total landings of a stock across fleets and metiers                
     \item{tdiscStock:} Total discard of a stock across fleets and metiers  
}
}
\examples{
\dontrun{
library(FLBEIA)
data(one)
landWStock.f(oneFl$fl1,"stk1")
landWStock(oneFl,"stk1")
}
}
