% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEco.R
\name{plotEco}
\alias{plotEco}
\title{Plots with fleets data}
\usage{
plotEco(obj, prob = c(0.95, 0.5, 0.05), pdfnm = "bc")
}
\arguments{
\item{obj}{An FLBEIA object.}

\item{pdfnm}{The name for the pdf document will be "Eco" and pdfnm separated by a line.}
}
\value{
A pdf with capacity, costs, effort, profits by fleet.
}
\description{
Return a pdf with plots using FLBEIA object (FLFleets and covars).
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)

data(res_flbeia)
plotEco(oneRes, pdfnm='one')
}
}
