% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFLBiols.R
\name{plotFLBiols}
\alias{plotFLBiols}
\title{Plots with biological data}
\usage{
plotFLBiols(biols, prob = c(0.95, 0.5, 0.05), pdfnm = "bc")
}
\arguments{
\item{biols}{A FLBiols object}

\item{prob}{a numeric vector with the probabilities used to calculate the quantiles.}

\item{pdfnm}{The name for the pdf document will be stock's name and pdfnm separated by a line.}
}
\value{
A pdf for each stock with plots.
}
\description{
For each stock, return a pdf with plots using FLBiols object.
}
\details{
\itemize{
     \item{Each pdf contains biomass in numbers at age, mean weight at age, fecundity, natural mortality, maturity, spawning, recruitment and spawning stock biomass} 
}
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(res_flbeia)
plotFLBiols(oneRes$biols, pdfname='oneRes')
}
}
