% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFLFleets.R
\name{plotFLFleets}
\alias{plotFLFleets}
\title{Plots with fleets data}
\usage{
plotFLFleets(fleets, prob = c(0.95, 0.5, 0.05), pdfnm = "bc")
}
\arguments{
\item{fleets}{A FLFleets object.}

\item{pdfnm}{The name for the pdf document will be the fleet's name and pdfnm separated by a line.}
}
\value{
A pdf for each fleet with plots.
}
\description{
For each fleet, return a pdf with plots using FLFleets object.
}
\details{
For each fleet, the pdf contains plots of:
\itemize{
     \item Catch, discards, landings, capacity, crewshare, 
     effort, fcost, effshare
     \item For each metier: landings and discards at age in numbers 
     and mean weight, alpha, beta and catch.q 
}
}
\examples{
\dontrun{
library(FLBEIA)
library(ggplot2)
data(res_flbeia)
plotFLFleets(oneFl,pdfnm='oneFl')
}
}
