% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Summary.R
\name{revenue_flbeia}
\alias{revenue_flbeia}
\alias{costs_flbeia}
\alias{totvcost_flbeia}
\alias{totfcost_flbeia}
\alias{taxcost_flbeia}
\alias{price_flbeia}
\title{Economic summary functions.}
\usage{
revenue_flbeia(fleet)

costs_flbeia(fleet, covars, flnm = NULL, fleet.ctrl = NULL, advice = NULL)

totvcost_flbeia(fleet, fleet.ctrl, advice, covars = NULL)

totfcost_flbeia(fleet, covars, flnm = NULL)

taxcost_flbeia(fleet, fleet.ctrl, advice)

price_flbeia(fleet, stock)
}
\arguments{
\item{fleet}{An element of FLfleets object.}

\item{covars}{List of FLQuants with information on covariates.}

\item{flnm}{Names of the fleets.}

\item{advice}{List of two FLquants, with TAC and quota share for all the stocks (only required if taxes = TRUE)}

\item{stock}{An FLStock object.}

\item{fleets.ctrl}{FLquant with quotas for all the stocks (only required if fleets.ctrl[[flnm]]$taxes == TRUE)}

\item{fleets.ctl}{FLquant with quotas for all the stocks (only required if taxes = TRUE)}
}
\description{
These functions provide summary results of costs, prices and revenues. Provided data can be dessagregated by fleet or by metier depending on the selected function.
}
\details{
\itemize{
      \item{revenue_flbeia}{ computes the revenue by fleet and metier. The revenue is computed as
       landings (weight) multiplied by the price.}
      \item{costs_flbeia}{ computes total costs as the sum of fixed and variable costs.}
      \item{totvcost_flbeia}{ computes the variable costs including crew share costs .}
      \item{totfcost_flbeia}{ computes the total costs by vessel.}
      \item{price_flbeia}{ computes the price by stock.} 
     }
}
