% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Summary.R
\name{revenue_flbeia}
\alias{revenue_flbeia}
\alias{costs_flbeia}
\alias{totvcost_flbeia}
\alias{totfcost_flbeia}
\alias{price_flbeia}
\title{Economic summary functions.}
\usage{
revenue_flbeia(fleet)

costs_flbeia(fleet, covars, flnm = NULL)

totvcost_flbeia(fleet)

totfcost_flbeia(fleet, covars, flnm = NULL)

price_flbeia(fleet, stock)
}
\arguments{
\item{fleet}{An element of FLfleets object.}

\item{covars}{A list of FLQuants used to store any kind of variables that are used within the simulation and are not stored in the standard objects.}

\item{flnm}{Names of the fleets.}

\item{stock}{An FLStock object.}
}
\description{
These functions provide summary results of costs, prices and revenues. Provided data can be dessagregated by fleet or by metier depending on the selected function.
}
\details{
\itemize{
      \item{revenue_flbeia}{ computes the revenue by fleet and metier. The revenue is computed as
       landings (weight) multiplied by the price.}
      \item{costs_flbeia}{ computes total costs as the sum of fixed and variable costs.}
      \item{totvcost_flbeia}{ computes the variable costs including crew share costs .}
      \item{totfcost_flbeia}{ computes the total costs by vessel.}
      \item{price_flbeia}{ computes the price by stock.} 
     }
}
