% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R,
%   R/Auxiliary_Functions_remove_units.R
\name{setUnitsNA}
\alias{setUnitsNA}
\alias{setUnitsNA-methods}
\alias{setUnitsNA,FLBiol-method}
\alias{setUnitsNA,FLBiols-method}
\alias{setUnitsNA,FLFleetExt-method}
\alias{setUnitsNA,FLFleetsExt-method}
\title{Method setUnitsNA}
\usage{
setUnitsNA(object)

\S4method{setUnitsNA}{FLBiol}(object)

\S4method{setUnitsNA}{FLBiols}(object)

\S4method{setUnitsNA}{FLFleetExt}(object)

\S4method{setUnitsNA}{FLFleetsExt}(object)
}
\arguments{
\item{object}{An object of class FLBiol, FLBiols, FLFleetExt or FLFleetsExt object.}
}
\value{
The same object with the units equal to NA in all the FLQuant slots.
}
\description{
Function to remove the units from the FLQuants of an object
}
\examples{

data(one)

st <- setUnitsNA(oneBio)
fl <- setUnitsNA(oneFl)

data(multi)

biols <- setUnitsNA(multiBio)
fleets <- setUnitsNA(multiFl)

data(res_flbeia)

res_biols  <- setUnitsNA(multiRes$biols)
res_fleets <- setUnitsNA(multiRes$fleets)

}
\keyword{methods}
\keyword{setUnitsNA}
