% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoSum_DAMARA_function.R, R/Results_Summary.R
\name{ecoSum_damara}
\alias{ecoSum_damara}
\alias{F_flbeia}
\alias{SSB_flbeia}
\alias{B_flbeia}
\alias{R_flbeia}
\alias{C_flbeia}
\alias{L_flbeia}
\alias{D_flbeia}
\alias{summary_flbeia}
\title{Biological summary functions}
\usage{
ecoSum_damara(fleets, flnms = "all", years, covars = NULL)

F_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

SSB_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

B_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

R_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

C_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

L_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

D_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

summary_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)
}
\arguments{
\item{fleets}{A FLFleetsExt object.}

\item{flnms}{Names of the fleet for which the indicators will be calculated.}

\item{years}{The  years for which the indicators are extracted.}

\item{covars}{List of FLQuants with information on covariates.}

\item{obj}{The output of the FLBEIA function.}
}
\value{
B_flbeia, F_flbeia... return an array with three dimensions (stock, year and iter).
The summary_flbeia function returns an array with 4 dimensions (stock, year, iter, indicator) with the value of all the indicators.
}
\description{
These functions return the biomass (B), fishing mortality (F),  spawning stock biomass (SSB), recruitment (R), catches (C), landings (L) and discards (D) indicators. Also indicators comparing the reference points with the actual values, Bpa, Blim, Btarget, Fpa, Flim and Ftarget, so the biomass indicators are TRUE if the biomass is above them, and the fishing mortality indicators are TRUE if the fishing mortality is below them. ssb2Btarget and f2Ftarget return the ratio between SSB and F and the target reference point.
}
\details{
\itemize{
      \item{B_flbeia}{ this function computes SSB.}
      \item{F_flbeia}{ this function computes fishing mortiality.}
      \item{SSB_flbeia}{ this function computes spawning stock biomass by species.}
      \item{R_flbeia}{ this function computes recruitment by stock. If the stock is defined by age this function the recruiment is computed. ;
                       If the stock is follows a biomass dynamics, this function gives the growth.}
      \item{C_flbeia}{ this function computes catches by fleets and stock.} 
      \item{L_flbeia}{ this function computes landings by fleets and stock.}
      \item{D_flbeia}{ this function computes the discards by fleets and stock.}
      \item{summary_flbeia}{ this function computes recruitment, SSB, fishing mortality,
                             total biomass for all stocks; 
                             and catch and landings for all fleets by year and stock.}
     }
}
